refactor(value): extract impl blocks from value/mod.rs into submodules (§7-8)
Reduce the 4398-line value/mod.rs by extracting impl blocks into ten
themed submodules under src/value/, all under 500 lines. mod.rs keeps all
type definitions (Value enum, *Data structs, aliases), free functions,
and mod/use declarations; only impl blocks move out. mod.rs drops to 1380
lines (still over 500, but it now holds only the cohesive data-model
definitions, which are one block).

- value_methods_{a,b,c}.rs (462/473/467) - impl Value, partitioned by theme
- value_lazy.rs (396)        - LazyList / LazyThunkData impls
- value_eq.rs (294)          - impl PartialEq for Value (Arc::ptr_eq logic, verbatim)
- value_async.rs (282)       - PromiseState / SharedPromise / SharedChannel impls
- value_instance.rs (256)    - AttrReadGuard / InstanceAttrs impls
- value_collections.rs (208) - ArrayKind / HashData / ArrayData impls
- value_setbagmix.rs (178)   - BagData / SetData / MixData impls
- value_enum.rs (29)         - impl EnumValue

Child modules see ancestor privates, so the split is relocation-only; two
private helpers (AttrReadGuard::new, InstanceAttrs::with_class) were bumped
to pub(super) for parent/sibling use. The impl multiset is byte-identical.
No behavior change.
