+++ src/value/mod.rs
@@ -242,65 +242,6 @@ pub(crate) struct BagData {
     pub declared_type: Option<String>,
 }
 
-impl BagData {
-    pub fn new(counts: HashMap<String, NumBigInt>) -> Self {
-        BagData {
-            counts,
-            original_keys: None,
-            value_type: None,
-            key_type: None,
-            declared_type: None,
-        }
-    }
-
-    pub fn with_original_keys(
-        counts: HashMap<String, NumBigInt>,
-        original_keys: HashMap<String, Value>,
-    ) -> Self {
-        BagData {
-            counts,
-            original_keys: Some(original_keys),
-            value_type: None,
-            key_type: None,
-            declared_type: None,
-        }
-    }
-
-    /// Whether container *type* metadata (element/key/declared type) is attached.
-    pub fn has_type_meta(&self) -> bool {
-        self.value_type.is_some() || self.key_type.is_some() || self.declared_type.is_some()
-    }
-
-    /// Get the original Value for a key, falling back to Str.
-    pub fn typed_key(&self, str_key: &str) -> Value {
-        if let Some(ref orig) = self.original_keys
-            && let Some(v) = orig.get(str_key)
-        {
-            return v.clone();
-        }
-        Value::Str(Arc::new(str_key.to_string()))
-    }
-}
-
-impl Deref for BagData {
-    type Target = HashMap<String, NumBigInt>;
-    fn deref(&self) -> &Self::Target {
-        &self.counts
-    }
-}
-
-impl DerefMut for BagData {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.counts
-    }
-}
-
-impl PartialEq for BagData {
-    fn eq(&self, other: &Self) -> bool {
-        self.counts == other.counts
-    }
-}
-
 /// Set data: wraps HashSet<String> with optional original-typed keys.
 /// Implements Deref to HashSet<String> so existing code works unchanged.
 #[derive(Debug, Clone)]
@@ -318,65 +259,6 @@ pub(crate) struct SetData {
     /// the container through copy-on-write and can never be inherited by an
     /// unrelated container via Arc-pointer reuse.
     pub declared_type: Option<String>,
-}
-
-impl SetData {
-    pub fn new(elements: HashSet<String>) -> Self {
-        SetData {
-            elements,
-            original_keys: None,
-            value_type: None,
-            key_type: None,
-            declared_type: None,
-        }
-    }
-
-    pub fn with_original_keys(
-        elements: HashSet<String>,
-        original_keys: HashMap<String, Value>,
-    ) -> Self {
-        SetData {
-            elements,
-            original_keys: Some(original_keys),
-            value_type: None,
-            key_type: None,
-            declared_type: None,
-        }
-    }
-
-    /// Whether container *type* metadata (element/key/declared type) is attached.
-    pub fn has_type_meta(&self) -> bool {
-        self.value_type.is_some() || self.key_type.is_some() || self.declared_type.is_some()
-    }
-
-    /// Get the original Value for a key, falling back to Str.
-    pub fn typed_key(&self, str_key: &str) -> Value {
-        if let Some(ref orig) = self.original_keys
-            && let Some(v) = orig.get(str_key)
-        {
-            return v.clone();
-        }
-        Value::Str(Arc::new(str_key.to_string()))
-    }
-}
-
-impl Deref for SetData {
-    type Target = HashSet<String>;
-    fn deref(&self) -> &Self::Target {
-        &self.elements
-    }
-}
-
-impl DerefMut for SetData {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.elements
-    }
-}
-
-impl PartialEq for SetData {
-    fn eq(&self, other: &Self) -> bool {
-        self.elements == other.elements
-    }
 }
 
 /// Mix data: wraps HashMap<String, f64> with optional original-typed keys.
@@ -398,65 +280,6 @@ pub(crate) struct MixData {
     pub declared_type: Option<String>,
 }
 
-impl MixData {
-    pub fn new(weights: HashMap<String, f64>) -> Self {
-        MixData {
-            weights,
-            original_keys: None,
-            value_type: None,
-            key_type: None,
-            declared_type: None,
-        }
-    }
-
-    pub fn with_original_keys(
-        weights: HashMap<String, f64>,
-        original_keys: HashMap<String, Value>,
-    ) -> Self {
-        MixData {
-            weights,
-            original_keys: Some(original_keys),
-            value_type: None,
-            key_type: None,
-            declared_type: None,
-        }
-    }
-
-    /// Whether container *type* metadata (element/key/declared type) is attached.
-    pub fn has_type_meta(&self) -> bool {
-        self.value_type.is_some() || self.key_type.is_some() || self.declared_type.is_some()
-    }
-
-    /// Get the original Value for a key, falling back to Str.
-    pub fn typed_key(&self, str_key: &str) -> Value {
-        if let Some(ref orig) = self.original_keys
-            && let Some(v) = orig.get(str_key)
-        {
-            return v.clone();
-        }
-        Value::Str(Arc::new(str_key.to_string()))
-    }
-}
-
-impl Deref for MixData {
-    type Target = HashMap<String, f64>;
-    fn deref(&self) -> &Self::Target {
-        &self.weights
-    }
-}
-
-impl DerefMut for MixData {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.weights
-    }
-}
-
-impl PartialEq for MixData {
-    fn eq(&self, other: &Self) -> bool {
-        self.weights == other.weights
-    }
-}
-
 mod aliased_mut;
 mod display;
 mod error;
@@ -468,6 +291,16 @@ pub(crate) mod types;
 pub(crate) mod types_eqv;
 pub(crate) mod types_isa;
 pub(crate) mod types_truthy;
+mod value_async;
+mod value_collections;
+mod value_enum;
+mod value_eq;
+mod value_instance;
+mod value_lazy;
+mod value_methods_a;
+mod value_methods_b;
+mod value_methods_c;
+mod value_setbagmix;
 pub(crate) use aliased_mut::arc_contents_mut;
 pub(crate) use types::what_type_name;
 
@@ -564,57 +397,6 @@ pub(crate) struct AttrReadGuard<'a> {
     addr: usize,
 }
 
-impl<'a> AttrReadGuard<'a> {
-    fn new(guard: std::sync::RwLockReadGuard<'a, HashMap<String, Value>>, addr: usize) -> Self {
-        HELD_READ_CELLS.with(|c| c.borrow_mut().push(addr));
-        Self {
-            guard: Some(guard),
-            addr,
-        }
-    }
-}
-
-impl std::ops::Deref for AttrReadGuard<'_> {
-    type Target = HashMap<String, Value>;
-    fn deref(&self) -> &Self::Target {
-        self.guard.as_ref().expect("attr read guard live")
-    }
-}
-
-impl Drop for AttrReadGuard<'_> {
-    fn drop(&mut self) {
-        let still_held = HELD_READ_CELLS.with(|c| {
-            let mut v = c.borrow_mut();
-            if let Some(pos) = v.iter().rposition(|&a| a == self.addr) {
-                v.swap_remove(pos);
-            }
-            v.contains(&self.addr)
-        });
-        // Release this read lock before flushing, so a deferred blocking write
-        // does not deadlock against it.
-        self.guard = None;
-        if still_held {
-            return;
-        }
-        let flush = PENDING_CELL_WRITES.with(|p| {
-            let mut v = p.borrow_mut();
-            let mut mine: Vec<(AttrCell, HashMap<String, Value>)> = Vec::new();
-            v.retain(|(a, cell, map)| {
-                if *a == self.addr {
-                    mine.push((cell.clone(), map.clone()));
-                    false
-                } else {
-                    true
-                }
-            });
-            mine
-        });
-        for (cell, map) in flush {
-            *write_attrs(&cell) = map;
-        }
-    }
-}
-
 #[derive(Debug)]
 pub(crate) struct InstanceAttrs {
     class_name: Symbol,
@@ -664,21 +446,6 @@ pub fn lookup_container_constraint(cell: &Arc<Mutex<Value>>) -> Option<String> {
         .cloned()
 }
 
-impl Clone for InstanceAttrs {
-    /// Deep, independent copy: a fresh cell with a snapshot of the map. Used for
-    /// `.clone`-style independent copies and `temp`/`let` snapshots; it must NOT
-    /// share the cell — sharing flows through `Arc<InstanceAttrs>`. The copy does
-    /// not participate in DESTROY refcounting (`queue_destroy = false`).
-    fn clone(&self) -> Self {
-        Self {
-            class_name: self.class_name,
-            attributes: Arc::new(RwLock::new(read_attrs(&self.attributes).clone())),
-            id: self.id,
-            queue_destroy: false,
-        }
-    }
-}
-
 /// Write `map` into `cell`, deferring if the current thread already holds a read
 /// guard on it.
 ///
@@ -715,192 +482,6 @@ fn write_attrs(
     cell: &RwLock<HashMap<String, Value>>,
 ) -> std::sync::RwLockWriteGuard<'_, HashMap<String, Value>> {
     cell.write().unwrap_or_else(|e| e.into_inner())
-}
-
-impl InstanceAttrs {
-    pub(crate) fn new(
-        class_name: Symbol,
-        attributes: HashMap<String, Value>,
-        id: u64,
-        queue_destroy: bool,
-    ) -> Self {
-        if queue_destroy && let Ok(mut counts) = live_instance_refcounts().lock() {
-            *counts.entry(id).or_insert(0) += 1;
-        }
-        let cell: AttrCell = Arc::new(RwLock::new(attributes));
-        Self {
-            class_name,
-            attributes: cell,
-            id,
-            queue_destroy,
-        }
-    }
-
-    /// Build an `InstanceAttrs` that shares an existing cell (used by the cell
-    /// reuse path in `make_instance_with_id`).
-    fn from_cell(class_name: Symbol, cell: AttrCell, id: u64, queue_destroy: bool) -> Self {
-        if queue_destroy && let Ok(mut counts) = live_instance_refcounts().lock() {
-            *counts.entry(id).or_insert(0) += 1;
-        }
-        Self {
-            class_name,
-            attributes: cell,
-            id,
-            queue_destroy,
-        }
-    }
-
-    // --- Attribute access API (Phase 3 — encapsulation boundary) ---
-    //
-    // The storage is a shared mutable cell (`Arc<RwLock<HashMap>>`). A locked
-    // cell cannot `Deref` to `&HashMap` (guard lifetime), so all access goes
-    // through these inherent methods. Reads take a read lock; mutations take a
-    // write lock and happen in place (visible to every alias).
-
-    /// Take a read lock over the attribute map. The guard derefs to `&HashMap`.
-    pub(crate) fn as_map(&self) -> AttrReadGuard<'_> {
-        read_attrs(&self.attributes)
-    }
-
-    /// An owned clone of the backing map.
-    pub(crate) fn to_map(&self) -> HashMap<String, Value> {
-        self.as_map().clone()
-    }
-
-    /// This instance's stable identity id.
-    pub(crate) fn instance_id(&self) -> u64 {
-        self.id
-    }
-
-    pub(crate) fn contains_key(&self, key: &str) -> bool {
-        self.as_map().contains_key(key)
-    }
-
-    /// In-place insert through the shared cell (visible to all aliases).
-    pub(crate) fn insert(&self, key: String, value: Value) -> Option<Value> {
-        write_attrs(&self.attributes).insert(key, value)
-    }
-
-    /// Mutate one attribute in place under the write lock, returning the
-    /// closure's result. Returns `None` if the key is absent. Replaces the old
-    /// `get_mut` (which cannot hand out a `&mut` past the guard).
-    pub(crate) fn with_attr_mut<R>(&self, key: &str, f: impl FnOnce(&mut Value) -> R) -> Option<R> {
-        let mut guard = write_attrs(&self.attributes);
-        guard.get_mut(key).map(f)
-    }
-
-    /// Insert `value` only if `key` is absent (the `entry(..).or_insert(..)`
-    /// idiom), in place under the write lock.
-    pub(crate) fn insert_if_absent(&self, key: String, value: Value) {
-        write_attrs(&self.attributes).entry(key).or_insert(value);
-    }
-
-    /// Phase 3 registry-removal: replace the whole attribute map in place through
-    /// this instance's shared cell, deadlock-safe with respect to a same-thread
-    /// read guard (see [`write_cell_respecting_reads`]). Because every alias of the
-    /// instance shares this `Arc<InstanceAttrs>` cell, the new map is visible
-    /// everywhere — in this frame, any caller frame, a `ContainerRef`-boxed
-    /// capture, a role `Mixin`, or a nested attribute of another instance. This is
-    /// the in-place replacement for the legacy id→cell registry writeback
-    /// (`overwrite_instance_bindings_by_identity` / `update_instance_cell`), which
-    /// computed an updated `HashMap` and looked the cell up by id.
-    pub(crate) fn commit_attrs(&self, map: HashMap<String, Value>) {
-        write_cell_respecting_reads(&self.attributes, map);
-    }
-
-    /// Phase 3 cell-CAS: atomically compare-and-swap one attribute under a
-    /// single write lock. When `matches(current)` returns true the new value is
-    /// stored; returns `(current, swapped)`. The cell's write lock is the
-    /// atomic primitive for cross-thread `cas`/atomic ops on instance
-    /// attributes — every alias of the instance shares this cell, so the swap
-    /// is immediately visible everywhere (no shared_vars side channel).
-    pub(crate) fn compare_and_swap(
-        &self,
-        key: &str,
-        matches: impl FnOnce(&Value) -> bool,
-        new: Value,
-    ) -> (Value, bool) {
-        let mut guard = write_attrs(&self.attributes);
-        let current = guard.get(key).cloned().unwrap_or(Value::Nil);
-        let swapped = matches(&current);
-        if swapped {
-            guard.insert(key.to_string(), new);
-        }
-        (current, swapped)
-    }
-
-    /// Phase 3 cell-CAS: atomically read-modify-write one attribute under a
-    /// single write lock (atomic add / increment / decrement). Returns
-    /// `(old, new)`; an error from `f` leaves the attribute unchanged.
-    #[allow(clippy::result_large_err)]
-    pub(crate) fn fetch_update(
-        &self,
-        key: &str,
-        f: impl FnOnce(&Value) -> Result<Value, RuntimeError>,
-    ) -> Result<(Value, Value), RuntimeError> {
-        let mut guard = write_attrs(&self.attributes);
-        let current = guard.get(key).cloned().unwrap_or(Value::Nil);
-        let next = f(&current)?;
-        guard.insert(key.to_string(), next.clone());
-        Ok((current, next))
-    }
-
-    /// Build an `InstanceAttrs` that SHARES this cell but carries a different
-    /// `class_name` (rebless / role mixin). The mutation visibility comes from the
-    /// shared cell; only the class tag differs.
-    fn with_class(&self, class_name: Symbol) -> Self {
-        Self::from_cell(
-            class_name,
-            Arc::clone(&self.attributes),
-            self.id,
-            self.queue_destroy,
-        )
-    }
-}
-
-impl PartialEq for InstanceAttrs {
-    fn eq(&self, other: &Self) -> bool {
-        if Arc::ptr_eq(&self.attributes, &other.attributes) {
-            return true;
-        }
-        *read_attrs(&self.attributes) == *read_attrs(&other.attributes)
-    }
-}
-
-impl Drop for InstanceAttrs {
-    fn drop(&mut self) {
-        if !self.queue_destroy {
-            return;
-        }
-        // Suppress recursive DESTROY queuing when we're already inside a DESTROY handler
-        if is_in_destroy_handler() {
-            return;
-        }
-        let should_queue = if let Ok(mut counts) = live_instance_refcounts().lock() {
-            match counts.get_mut(&self.id) {
-                Some(count) if *count > 1 => {
-                    *count -= 1;
-                    false
-                }
-                Some(_) => {
-                    counts.remove(&self.id);
-                    true
-                }
-                None => true,
-            }
-        } else {
-            true
-        };
-        if !should_queue {
-            return;
-        }
-        let _ = PENDING_INSTANCE_DESTROYS.try_with(|pending| {
-            pending.borrow_mut().push(PendingInstanceDestroy {
-                class_name: self.class_name,
-                attributes: read_attrs(&self.attributes).clone(),
-            });
-        });
-    }
 }
 
 pub(crate) fn next_instance_id() -> u64 {
@@ -1227,46 +808,6 @@ pub enum ArrayKind {
     Lazy,
 }
 
-impl ArrayKind {
-    /// True for `Array`, `ItemArray`, `Shaped`, and `Lazy` (the `[...]` constructor or shaped declaration).
-    pub fn is_real_array(self) -> bool {
-        matches!(
-            self,
-            ArrayKind::Array | ArrayKind::ItemArray | ArrayKind::Shaped | ArrayKind::Lazy
-        )
-    }
-
-    /// True for lazy arrays backed by an infinite source.
-    pub fn is_lazy(self) -> bool {
-        matches!(self, ArrayKind::Lazy)
-    }
-
-    /// True for `ItemList` and `ItemArray` (Scalar-wrapped).
-    pub fn is_itemized(self) -> bool {
-        matches!(self, ArrayKind::ItemList | ArrayKind::ItemArray)
-    }
-
-    /// Wrap in a Scalar container (`.item`).
-    pub fn itemize(self) -> Self {
-        match self {
-            ArrayKind::List => ArrayKind::ItemList,
-            ArrayKind::Array => ArrayKind::ItemArray,
-            other => other,
-        }
-    }
-
-    /// Remove Scalar wrapper (decontainerize).
-    /// Strips itemization only: `ItemArray → Array`, `ItemList → List`.
-    /// Non-itemized kinds are unchanged: `Array → Array`, `List → List`.
-    pub fn decontainerize(self) -> Self {
-        match self {
-            ArrayKind::ItemArray => ArrayKind::Array,
-            ArrayKind::ItemList => ArrayKind::List,
-            other => other,
-        }
-    }
-}
-
 /// Backing storage for `Value::Hash`. Bundles the key/value map with its
 /// container type metadata (element/key types) and object-hash original keys,
 /// so the metadata travels WITH the hash through copy-on-write and rebuilds —
@@ -1307,114 +848,6 @@ pub struct HashData {
     pub itemized: bool,
 }
 
-impl HashData {
-    pub fn new(map: HashMap<String, Value>) -> Self {
-        HashData {
-            map,
-            value_type: None,
-            key_type: None,
-            declared_type: None,
-            original_keys: None,
-            default: None,
-            itemized: false,
-        }
-    }
-
-    /// Whether any container metadata is attached (type or object-hash keys).
-    pub fn has_meta(&self) -> bool {
-        self.has_type_meta() || self.original_keys.is_some()
-    }
-
-    /// Whether container *type* metadata (element/key/declared type) is attached.
-    /// This is the authoritative replacement for the `hash_type_metadata` side
-    /// table: a freshly-built hash literal has `false` here, so it can never
-    /// inherit a stale typed-hash entry via Arc-pointer reuse.
-    pub fn has_type_meta(&self) -> bool {
-        self.value_type.is_some() || self.key_type.is_some() || self.declared_type.is_some()
-    }
-
-    /// Clear all container *type* metadata (used when re-tagging in place).
-    pub fn clear_type_meta(&mut self) {
-        self.value_type = None;
-        self.key_type = None;
-        self.declared_type = None;
-    }
-
-    /// Whether reading this hash's entries should yield typed (original) keys
-    /// rather than plain `Str` keys. Only object hashes (`my %h{KeyType}`,
-    /// marked by `key_type`) and hashes coerced from a Set/Bag/Mix (tagged with
-    /// the `__mutsu_setty_origin` marker) preserve typed keys. A *plain* hash
-    /// always reports `Str` keys even if construction speculatively recorded a
-    /// typed key (e.g. the `Int` 1 from `my %h = 1..6`), because Raku hash keys
-    /// are always stringified.
-    pub fn has_typed_keys(&self) -> bool {
-        self.original_keys.as_ref().is_some_and(|orig| {
-            !orig.is_empty()
-                && (self.key_type.is_some() || orig.contains_key("__mutsu_setty_origin"))
-        })
-    }
-
-    /// Get the original (typed) key Value for a stored string key. For object
-    /// hashes (`my %h{Int}`) the stored key is a `.WHICH` string (e.g.
-    /// `"Int|1"`); this returns the real key object (`Int(1)`). Plain hashes
-    /// report a `Str`. Mirrors `BagData::typed_key`. Honors [`Self::has_typed_keys`]:
-    /// a plain hash's speculative `original_keys` is ignored.
-    pub fn typed_key(&self, str_key: &str) -> Value {
-        if self.has_typed_keys()
-            && let Some(ref orig) = self.original_keys
-            && let Some(v) = orig.get(str_key)
-        {
-            return v.clone();
-        }
-        Value::Str(Arc::new(str_key.to_string()))
-    }
-
-    /// Build a `Pair` for a hash entry, using the original typed key for object
-    /// hashes. Plain string keys yield `Value::Pair`; non-string typed keys
-    /// (object hashes) yield `Value::ValuePair`. Behaviour-neutral for plain
-    /// hashes (returns `Value::Pair(str_key, value)` as before).
-    /// Build a `Pair` for a hash entry `(str_key, value)`, honoring object-hash
-    /// typed keys. The value is decontainerized: a hash element is stored as a
-    /// `ContainerRef` cell, but the pair must carry the inner value (matching a
-    /// `%h<k>` read and `.values`) — otherwise iterating a hash as pairs leaks
-    /// the cell and `+`/`.elems` on the pair value misbehave (the cell counts as
-    /// a single scalar item). See t/bind-hash-value-pairs.t.
-    pub fn typed_pair(&self, str_key: &str, value: Value) -> Value {
-        let value = value.deref_container();
-        match self.typed_key(str_key) {
-            Value::Str(s) => Value::Pair((*s).clone(), Box::new(value)),
-            other => Value::ValuePair(Box::new(other), Box::new(value)),
-        }
-    }
-}
-
-impl std::ops::Deref for HashData {
-    type Target = HashMap<String, Value>;
-    fn deref(&self) -> &HashMap<String, Value> {
-        &self.map
-    }
-}
-
-impl std::ops::DerefMut for HashData {
-    fn deref_mut(&mut self) -> &mut HashMap<String, Value> {
-        &mut self.map
-    }
-}
-
-impl From<HashMap<String, Value>> for HashData {
-    fn from(map: HashMap<String, Value>) -> Self {
-        HashData::new(map)
-    }
-}
-
-/// Hash equality ignores container metadata — only the key/value map matters
-/// (preserves the prior `Arc<HashMap>` PartialEq semantics).
-impl PartialEq for HashData {
-    fn eq(&self, other: &Self) -> bool {
-        self.map == other.map
-    }
-}
-
 /// Backing data for `Value::Array`: the element vector plus embedded
 /// container type metadata (mirrors `HashData`). `Deref`/`DerefMut` to
 /// `Vec<Value>` keep the overwhelming majority of read sites
@@ -1445,65 +878,6 @@ pub struct ArrayData {
     pub shape: Option<Vec<usize>>,
 }
 
-impl ArrayData {
-    pub fn new(items: Vec<Value>) -> Self {
-        ArrayData {
-            items,
-            value_type: None,
-            key_type: None,
-            declared_type: None,
-            default: None,
-            shape: None,
-        }
-    }
-
-    /// Whether container *type* metadata (element/key/declared type) is attached.
-    pub fn has_type_meta(&self) -> bool {
-        self.value_type.is_some() || self.key_type.is_some() || self.declared_type.is_some()
-    }
-}
-
-impl std::ops::Deref for ArrayData {
-    type Target = Vec<Value>;
-    fn deref(&self) -> &Vec<Value> {
-        &self.items
-    }
-}
-
-impl std::ops::DerefMut for ArrayData {
-    fn deref_mut(&mut self) -> &mut Vec<Value> {
-        &mut self.items
-    }
-}
-
-impl From<Vec<Value>> for ArrayData {
-    fn from(items: Vec<Value>) -> Self {
-        ArrayData::new(items)
-    }
-}
-
-/// Array equality ignores container metadata — only the elements matter
-/// (preserves the prior `Arc<Vec<Value>>` PartialEq semantics).
-impl PartialEq for ArrayData {
-    fn eq(&self, other: &Self) -> bool {
-        self.items == other.items
-    }
-}
-
-impl FromIterator<Value> for ArrayData {
-    fn from_iter<I: IntoIterator<Item = Value>>(iter: I) -> Self {
-        ArrayData::new(iter.into_iter().collect())
-    }
-}
-
-impl<'a> IntoIterator for &'a ArrayData {
-    type Item = &'a Value;
-    type IntoIter = std::slice::Iter<'a, Value>;
-    fn into_iter(self) -> Self::IntoIter {
-        self.items.iter()
-    }
-}
-
 /// Value stored in an enum variant: an integer, a string, or an arbitrary Value.
 #[derive(Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
 pub enum EnumValue {
@@ -1511,34 +885,6 @@ pub enum EnumValue {
     Str(String),
     /// Arbitrary value (e.g., Array for typed enums like `my Array enum ...`)
     Generic(Box<Value>),
-}
-
-impl EnumValue {
-    /// Return the integer value, or 0 for string/generic enums.
-    pub fn as_i64(&self) -> i64 {
-        match self {
-            EnumValue::Int(i) => *i,
-            EnumValue::Str(_) | EnumValue::Generic(_) => 0,
-        }
-    }
-
-    /// Return the string representation for `.Str` coercion.
-    pub fn to_str_value(&self) -> String {
-        match self {
-            EnumValue::Int(i) => i.to_string(),
-            EnumValue::Str(s) => s.clone(),
-            EnumValue::Generic(v) => v.to_str_context(),
-        }
-    }
-
-    /// Convert to a runtime Value.
-    pub fn to_value(&self) -> Value {
-        match self {
-            EnumValue::Int(i) => Value::Int(*i),
-            EnumValue::Str(s) => Value::str(s.clone()),
-            EnumValue::Generic(v) => v.as_ref().clone(),
-        }
-    }
 }
 
 #[allow(private_interfaces)]
@@ -1917,47 +1263,6 @@ pub(crate) struct LazyList {
     pub(crate) closure_seq: Option<Mutex<ClosureSeqState>>,
 }
 
-impl std::fmt::Debug for LazyList {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("LazyList")
-            .field("body_len", &self.body.len())
-            .field("has_compiled_code", &self.compiled_code.is_some())
-            .field("has_coroutine", &self.coroutine.is_some())
-            .field("has_lazy_pipe", &self.lazy_pipe.is_some())
-            .finish()
-    }
-}
-
-impl Clone for LazyList {
-    fn clone(&self) -> Self {
-        Self {
-            body: self.body.clone(),
-            env: self.env.clone(),
-            cache: Mutex::new(self.cache.lock().unwrap().clone()),
-            compiled_code: self.compiled_code.clone(),
-            compiled_fns: self.compiled_fns.clone(),
-            elems_count: self.elems_count.clone(),
-            scan_spec: self
-                .scan_spec
-                .as_ref()
-                .map(|s| Mutex::new(s.lock().unwrap().clone())),
-            sequence_spec: self.sequence_spec.clone(),
-            coroutine: self
-                .coroutine
-                .as_ref()
-                .map(|c| Mutex::new(c.lock().unwrap().clone())),
-            lazy_pipe: self
-                .lazy_pipe
-                .as_ref()
-                .map(|p| Mutex::new(p.lock().unwrap().clone())),
-            closure_seq: self
-                .closure_seq
-                .as_ref()
-                .map(|c| Mutex::new(c.lock().unwrap().clone())),
-        }
-    }
-}
-
 /// Placeholder string rendered for a genuinely-lazy list under
 /// gist/Str/raku/perl, matching Rakudo: `...` for `.Str`, `[...]` for a list
 /// held in `@` array context, `(...)` for a bare Seq.
@@ -1971,351 +1276,6 @@ pub(crate) fn lazy_list_placeholder(method: &str, array_context: bool) -> String
     }
 }
 
-impl LazyList {
-    /// Marker key inserted into `env` when a lazy list is bound/assigned into
-    /// an `@` array slot. A `Value::LazyList` carries no `[`-vs-`(` context on
-    /// its own, so this flag lets gist/`.WHAT` render `[...]`/`Array` (held in
-    /// `@a`) instead of `(...)`/`Seq` (a bare `$s` Seq).
-    pub(crate) const ARRAY_CONTEXT_MARKER: &'static str = "__mutsu_lazylist_array_context";
-
-    /// Whether this lazy list was assigned into an `@` array (see marker above).
-    pub(crate) fn in_array_context(&self) -> bool {
-        matches!(
-            self.env.get(Self::ARRAY_CONTEXT_MARKER),
-            Some(Value::Bool(true))
-        )
-    }
-
-    /// True when this list is genuinely lazy (`.is-lazy`), so gist/Str/raku
-    /// render a placeholder instead of materializing it.
-    ///
-    /// An infinite sequence/closure/scan/map-grep generator is only ever stored
-    /// as a *live* `LazyList` when actually infinite (finite ones materialize to
-    /// a `Seq`), so those specs are unconditionally lazy. A gather coroutine (or
-    /// unevaluated body), however, is lazy **only** when explicitly marked
-    /// `lazy` — a plain `gather` is `.is-lazy` `False` in Rakudo and must
-    /// materialize on gist/Str rather than render a placeholder.
-    pub(crate) fn is_genuinely_lazy(&self) -> bool {
-        if self.sequence_spec.is_some()
-            || self.lazy_pipe.is_some()
-            || self.closure_seq.is_some()
-            || self.scan_spec.is_some()
-        {
-            return true;
-        }
-        (self.coroutine.is_some() || !self.body.is_empty() || self.compiled_code.is_some())
-            && self.is_lazy_marked()
-    }
-
-    /// Whether this list was produced from a `gather` block (carries the
-    /// `__mutsu_lazylist_from_gather` env marker).
-    pub(crate) fn is_from_gather(&self) -> bool {
-        matches!(
-            self.env.get("__mutsu_lazylist_from_gather"),
-            Some(Value::Bool(true))
-        )
-    }
-
-    /// Whether this list is an infinite arithmetic/geometric sequence
-    /// (`1..*`, `1,2,3...*`) or an infinite closure sequence (`1,1,*+*...*`).
-    /// These reify on demand via `force_lazy_list_vm_n` (`extend_sequence_cache`
-    /// / `extend_closure_sequence`), so a method that needs the whole list must
-    /// raise `X::Cannot::Lazy` rather than read the (tiny) seed cache (L2b).
-    pub(crate) fn is_infinite_spec(&self) -> bool {
-        self.sequence_spec.is_some() || self.closure_seq.is_some()
-    }
-
-    /// Gate for the VM force/incremental-pull dispatch block: a gather-sourced
-    /// list (eager or `lazy`) or an infinite sequence/closure spec.
-    pub(crate) fn needs_vm_lazy_dispatch(&self) -> bool {
-        self.is_from_gather() || self.is_infinite_spec()
-    }
-
-    /// Whether this list carries the `lazy` prefix marker (set by the `lazy`
-    /// statement prefix / `.lazy` method).
-    fn is_lazy_marked(&self) -> bool {
-        matches!(
-            self.env.get("__mutsu_preserve_lazy_on_array_assign"),
-            Some(Value::Bool(true))
-        )
-    }
-
-    /// Return a clone of this list tagged as living in `@` array context.
-    pub(crate) fn with_array_context(&self) -> Self {
-        let mut cloned = self.clone();
-        cloned
-            .env
-            .insert(Self::ARRAY_CONTEXT_MARKER.to_string(), Value::Bool(true));
-        cloned
-    }
-
-    pub(crate) const LIST_CONTEXT_MARKER: &'static str = "__mutsu_lazylist_list_context";
-
-    /// Whether this lazy list was coerced via `.List` (so `.WHAT` is `List`,
-    /// not the default `Seq`). Mutually exclusive with array context in practice.
-    pub(crate) fn in_list_context(&self) -> bool {
-        matches!(
-            self.env.get(Self::LIST_CONTEXT_MARKER),
-            Some(Value::Bool(true))
-        )
-    }
-
-    /// Return a clone of this list tagged as a `.List`-coerced list. Preserves
-    /// laziness (the generator is untouched) while making `.WHAT` report `List`.
-    pub(crate) fn with_list_context(&self) -> Self {
-        let mut cloned = self.clone();
-        cloned
-            .env
-            .insert(Self::LIST_CONTEXT_MARKER.to_string(), Value::Bool(true));
-        cloned
-    }
-
-    /// Create a pre-cached lazy list (no body to evaluate).
-    pub(crate) fn new_cached(items: Vec<Value>) -> Self {
-        Self {
-            body: Vec::new(),
-            env: crate::env::Env::new(),
-            cache: Mutex::new(Some(items)),
-            compiled_code: None,
-            compiled_fns: None,
-            elems_count: None,
-            scan_spec: None,
-            sequence_spec: None,
-            coroutine: None,
-            lazy_pipe: None,
-            closure_seq: None,
-        }
-    }
-
-    /// Create an infinite sequence lazy list that can generate elements on demand.
-    pub(crate) fn new_sequence(seeds: Vec<Value>, spec: SequenceSpec) -> Self {
-        Self {
-            body: Vec::new(),
-            env: crate::env::Env::new(),
-            cache: Mutex::new(Some(seeds)),
-            compiled_code: None,
-            compiled_fns: None,
-            elems_count: None,
-            scan_spec: None,
-            sequence_spec: Some(spec),
-            coroutine: None,
-            lazy_pipe: None,
-            closure_seq: None,
-        }
-    }
-
-    /// Create a lazy scan (triangle reduce) list that computes elements on demand.
-    pub(crate) fn new_scan(spec: ScanSpec) -> Self {
-        Self {
-            body: Vec::new(),
-            env: crate::env::Env::new(),
-            cache: Mutex::new(Some(Vec::new())),
-            compiled_code: None,
-            compiled_fns: None,
-            elems_count: None,
-            scan_spec: Some(Mutex::new(spec)),
-            sequence_spec: None,
-            coroutine: None,
-            lazy_pipe: None,
-            closure_seq: None,
-        }
-    }
-
-    /// Create a lazy `map`/`grep` pipeline stage over `source`.
-    ///
-    /// The result stays lazy: its elements are produced on demand by pulling
-    /// from `source` and applying `func`. The `__mutsu_lazylist_from_gather`
-    /// marker is set so the VM's `.head`/`.first`/index dispatch routes through
-    /// the bounded incremental-pull path.
-    pub(crate) fn new_pipe(source: Value, func: Value, is_grep: bool) -> Self {
-        let mut env = crate::env::Env::new();
-        env.insert(
-            "__mutsu_lazylist_from_gather".to_string(),
-            Value::Bool(true),
-        );
-        Self {
-            body: Vec::new(),
-            env,
-            cache: Mutex::new(Some(Vec::new())),
-            compiled_code: None,
-            compiled_fns: None,
-            elems_count: None,
-            scan_spec: None,
-            sequence_spec: None,
-            coroutine: None,
-            lazy_pipe: Some(Mutex::new(MapGrepSpec {
-                source,
-                func,
-                is_grep,
-                source_idx: 0,
-                done: false,
-                index_transform: None,
-            })),
-            closure_seq: None,
-        }
-    }
-
-    /// Create a lazy `.pairs`/`.antipairs`/`.kv` stage over `source`.
-    ///
-    /// Stays lazy (carries the gather + preserve markers so array assignment
-    /// keeps it lazy, matching Rakudo where these methods are `.is-lazy` over a
-    /// lazy list). Elements are produced on demand by pulling from `source` and
-    /// applying the index transform with the source position as the key.
-    pub(crate) fn new_index_pipe(source: Value, transform: IndexTransform) -> Self {
-        let mut env = crate::env::Env::new();
-        env.insert(
-            "__mutsu_lazylist_from_gather".to_string(),
-            Value::Bool(true),
-        );
-        env.insert(
-            "__mutsu_preserve_lazy_on_array_assign".to_string(),
-            Value::Bool(true),
-        );
-        Self {
-            body: Vec::new(),
-            env,
-            cache: Mutex::new(Some(Vec::new())),
-            compiled_code: None,
-            compiled_fns: None,
-            elems_count: None,
-            scan_spec: None,
-            sequence_spec: None,
-            coroutine: None,
-            lazy_pipe: Some(Mutex::new(MapGrepSpec {
-                source,
-                func: Value::Nil,
-                is_grep: false,
-                source_idx: 0,
-                done: false,
-                index_transform: Some(transform),
-            })),
-            closure_seq: None,
-        }
-    }
-
-    /// Create an infinite closure-based sequence (`1, 1, * + * ... *`).
-    ///
-    /// `seeds` is the initial element history (already includes any eagerly
-    /// generated prefix); `state` carries the generator closure so more
-    /// elements can be produced on demand via the VM.
-    pub(crate) fn new_closure_sequence(seeds: Vec<Value>, state: ClosureSeqState) -> Self {
-        Self {
-            body: Vec::new(),
-            env: crate::env::Env::new(),
-            cache: Mutex::new(Some(seeds)),
-            compiled_code: None,
-            compiled_fns: None,
-            elems_count: None,
-            scan_spec: None,
-            sequence_spec: None,
-            coroutine: None,
-            lazy_pipe: None,
-            closure_seq: Some(Mutex::new(state)),
-        }
-    }
-
-    /// Force a scan-based lazy list to compute up to `needed` elements.
-    /// Uses builtin arithmetic for common operators. Returns the cached elements.
-    /// This can be called from contexts without VM access (builtins, interpreter).
-    pub(crate) fn force_scan_to(&self, needed: usize) -> Vec<Value> {
-        let scan_mutex = match &self.scan_spec {
-            Some(s) => s,
-            None => return self.cache.lock().unwrap().clone().unwrap_or_default(),
-        };
-
-        let mut spec = scan_mutex.lock().unwrap();
-        let mut cache_guard = self.cache.lock().unwrap();
-        let out = cache_guard.get_or_insert_with(Vec::new);
-
-        if out.len() >= needed {
-            return out[..needed].to_vec();
-        }
-
-        let remaining = needed - out.len();
-        let already = spec.computed_count;
-        let source = spec.source.clone();
-        let base_op = spec.op.clone();
-        let negate = spec.negate;
-
-        // Generate source values
-        let new_values: Vec<Value> = match &source {
-            Value::Range(a, b) => {
-                let start = *a + already as i64;
-                let end = if *b == i64::MAX {
-                    *a + needed as i64
-                } else {
-                    *b
-                };
-                (start..=end).take(remaining).map(Value::Int).collect()
-            }
-            Value::RangeExcl(a, b) => {
-                let start = *a + already as i64;
-                let end = if *b == i64::MAX {
-                    *a + needed as i64
-                } else {
-                    *b
-                };
-                (start..end).take(remaining).map(Value::Int).collect()
-            }
-            _ => {
-                let items = crate::runtime::utils::value_to_list(&source);
-                items.into_iter().skip(already).take(remaining).collect()
-            }
-        };
-
-        let mut acc = spec.accumulator.clone();
-        for val in new_values {
-            acc = Some(match acc.take() {
-                None => {
-                    out.push(val.clone());
-                    val
-                }
-                Some(prev) => {
-                    let v = Self::scan_binary_op(&base_op, prev, val);
-                    let v = if negate { Value::Bool(!v.truthy()) } else { v };
-                    out.push(v.clone());
-                    v
-                }
-            });
-            spec.computed_count += 1;
-        }
-        spec.accumulator = acc;
-        out.clone()
-    }
-
-    /// Apply a binary operator for scan reduction. Supports common builtin ops.
-    fn scan_binary_op(op: &str, left: Value, right: Value) -> Value {
-        match op {
-            "+" => crate::builtins::arith::arith_add(left, right).unwrap_or(Value::Nil),
-            "-" => crate::builtins::arith::arith_sub(left, right),
-            "*" => crate::builtins::arith::arith_mul(left, right),
-            "/" => crate::builtins::arith::arith_div(left, right).unwrap_or(Value::Nil),
-            "%" | "mod" => crate::builtins::arith::arith_mod(left, right).unwrap_or(Value::Nil),
-            "**" => crate::builtins::arith::arith_pow(left, right),
-            "~" => Value::str(format!(
-                "{}{}",
-                left.to_string_value(),
-                right.to_string_value()
-            )),
-            "max" => {
-                if left.to_f64() >= right.to_f64() {
-                    left
-                } else {
-                    right
-                }
-            }
-            "min" => {
-                if left.to_f64() <= right.to_f64() {
-                    left
-                } else {
-                    right
-                }
-            }
-            _ => Value::Nil, // Unsupported op — VM path handles these
-        }
-    }
-}
-
 // --- LazyThunkData: deferred evaluation with caching ---
 
 #[derive(Debug)]
@@ -2324,15 +1284,6 @@ pub(crate) struct LazyThunkData {
     pub(crate) thunk: Value,
     /// Cached result: None means not yet evaluated.
     pub(crate) cache: Mutex<Option<Value>>,
-}
-
-impl Clone for LazyThunkData {
-    fn clone(&self) -> Self {
-        Self {
-            thunk: self.thunk.clone(),
-            cache: Mutex::new(self.cache.lock().unwrap().clone()),
-        }
-    }
 }
 
 // --- SharedPromise: thread-safe promise state ---
@@ -2349,169 +1300,9 @@ struct PromiseState {
     thread_payload: Option<Box<dyn std::any::Any + Send>>,
 }
 
-impl std::fmt::Debug for PromiseState {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("PromiseState")
-            .field("status", &self.status)
-            .field("result", &self.result)
-            .field("output", &self.output)
-            .field("stderr_output", &self.stderr_output)
-            .field("class_name", &self.class_name)
-            .finish()
-    }
-}
-
 #[derive(Debug, Clone)]
 pub(crate) struct SharedPromise {
     inner: Arc<(Mutex<PromiseState>, Condvar)>,
-}
-
-impl SharedPromise {
-    /// Returns a raw pointer to the inner Arc data, for use in WHICH identity.
-    pub fn arc_ptr(&self) -> *const () {
-        Arc::as_ptr(&self.inner) as *const ()
-    }
-
-    pub(crate) fn new() -> Self {
-        Self::new_with_class(Symbol::intern("Promise"))
-    }
-
-    pub(crate) fn new_with_class(class_name: Symbol) -> Self {
-        Self {
-            inner: Arc::new((
-                Mutex::new(PromiseState {
-                    status: "Planned".to_string(),
-                    result: Value::Nil,
-                    output: String::new(),
-                    stderr_output: String::new(),
-                    class_name,
-                    thread_payload: None,
-                }),
-                Condvar::new(),
-            )),
-        }
-    }
-
-    #[allow(dead_code)]
-    pub(crate) fn new_kept(result: Value) -> Self {
-        Self {
-            inner: Arc::new((
-                Mutex::new(PromiseState {
-                    status: "Kept".to_string(),
-                    result,
-                    output: String::new(),
-                    stderr_output: String::new(),
-                    class_name: Symbol::intern("Promise"),
-                    thread_payload: None,
-                }),
-                Condvar::new(),
-            )),
-        }
-    }
-
-    pub(crate) fn class_name(&self) -> Symbol {
-        let (lock, _) = &*self.inner;
-        lock.lock().unwrap().class_name
-    }
-
-    pub(crate) fn id(&self) -> usize {
-        Arc::as_ptr(&self.inner) as usize
-    }
-
-    /// Store an opaque payload to be handed off to whoever awaits the
-    /// promise. Must be called before `keep`/`break_with` so the awaiter
-    /// can observe it via `take_thread_payload`.
-    pub(crate) fn set_thread_payload(&self, payload: Box<dyn std::any::Any + Send>) {
-        let (lock, _) = &*self.inner;
-        let mut state = lock.lock().unwrap();
-        state.thread_payload = Some(payload);
-    }
-
-    pub(crate) fn take_thread_payload(&self) -> Option<Box<dyn std::any::Any + Send>> {
-        let (lock, _) = &*self.inner;
-        let mut state = lock.lock().unwrap();
-        state.thread_payload.take()
-    }
-
-    pub(crate) fn keep(&self, result: Value, output: String, stderr: String) {
-        let (lock, cvar) = &*self.inner;
-        let mut state = lock.lock().unwrap();
-        state.status = "Kept".to_string();
-        state.result = result;
-        state.output = output;
-        state.stderr_output = stderr;
-        cvar.notify_all();
-    }
-
-    /// Try to keep; returns Err(current_status) if already kept/broken.
-    pub(crate) fn try_keep(&self, result: Value) -> Result<(), String> {
-        let (lock, cvar) = &*self.inner;
-        let mut state = lock.lock().unwrap();
-        if state.status != "Planned" {
-            return Err(state.status.clone());
-        }
-        state.status = "Kept".to_string();
-        state.result = result;
-        cvar.notify_all();
-        Ok(())
-    }
-
-    pub(crate) fn break_with(&self, error: Value, output: String, stderr: String) {
-        let (lock, cvar) = &*self.inner;
-        let mut state = lock.lock().unwrap();
-        state.status = "Broken".to_string();
-        state.result = error;
-        state.output = output;
-        state.stderr_output = stderr;
-        cvar.notify_all();
-    }
-
-    /// Try to break; returns Err(current_status) if already kept/broken.
-    pub(crate) fn try_break(&self, error: Value) -> Result<(), String> {
-        let (lock, cvar) = &*self.inner;
-        let mut state = lock.lock().unwrap();
-        if state.status != "Planned" {
-            return Err(state.status.clone());
-        }
-        state.status = "Broken".to_string();
-        state.result = error;
-        cvar.notify_all();
-        Ok(())
-    }
-
-    /// Check if promise is resolved (Kept or Broken).
-    pub(crate) fn is_resolved(&self) -> bool {
-        let (lock, _) = &*self.inner;
-        lock.lock().unwrap().status != "Planned"
-    }
-
-    pub(crate) fn status(&self) -> String {
-        let (lock, _) = &*self.inner;
-        lock.lock().unwrap().status.clone()
-    }
-
-    pub(crate) fn wait(&self) -> (Value, String, String) {
-        let (lock, cvar) = &*self.inner;
-        let mut state = lock.lock().unwrap();
-        while state.status == "Planned" {
-            state = cvar.wait(state).unwrap();
-        }
-        (
-            state.result.clone(),
-            state.output.clone(),
-            state.stderr_output.clone(),
-        )
-    }
-
-    pub(crate) fn result_blocking(&self) -> Value {
-        self.wait().0
-    }
-}
-
-impl PartialEq for SharedPromise {
-    fn eq(&self, other: &Self) -> bool {
-        Arc::ptr_eq(&self.inner, &other.inner)
-    }
 }
 
 // --- SharedChannel: thread-safe channel ---
@@ -2529,1815 +1320,6 @@ struct ChannelState {
 #[derive(Debug, Clone)]
 pub(crate) struct SharedChannel {
     inner: Arc<(Mutex<ChannelState>, Condvar)>,
-}
-
-impl SharedChannel {
-    /// Returns a raw pointer to the inner Arc data, for use in WHICH identity.
-    pub fn arc_ptr(&self) -> *const () {
-        Arc::as_ptr(&self.inner) as *const ()
-    }
-
-    pub(crate) fn new() -> Self {
-        let closed_promise = SharedPromise::new();
-        Self {
-            inner: Arc::new((
-                Mutex::new(ChannelState {
-                    queue: std::collections::VecDeque::new(),
-                    send_closed: false,
-                    drained_closed: false,
-                    failure: None,
-                    closed_promise,
-                    supplier_ids: Vec::new(),
-                }),
-                Condvar::new(),
-            )),
-        }
-    }
-
-    pub(crate) fn send(&self, value: Value) {
-        let (lock, cvar) = &*self.inner;
-        let mut state = lock.lock().unwrap();
-        if state.send_closed {
-            return;
-        }
-        state.queue.push_back(value);
-        cvar.notify_one();
-    }
-
-    fn finish_if_drained(state: &mut ChannelState) {
-        if state.send_closed && state.queue.is_empty() && !state.drained_closed {
-            state.drained_closed = true;
-            if let Some(err) = state.failure.clone() {
-                let _ = state.closed_promise.try_break(err);
-            } else {
-                let _ = state.closed_promise.try_keep(Value::Nil);
-            }
-        }
-    }
-
-    pub(crate) fn receive_result(&self) -> Result<Value, Value> {
-        let (lock, cvar) = &*self.inner;
-        let mut state = lock.lock().unwrap();
-        loop {
-            if let Some(val) = state.queue.pop_front() {
-                Self::finish_if_drained(&mut state);
-                return Ok(val);
-            }
-            if state.drained_closed {
-                if let Some(err) = state.failure.clone() {
-                    return Err(err);
-                }
-                return Ok(Value::Nil);
-            }
-            state = cvar.wait(state).unwrap();
-        }
-    }
-
-    pub(crate) fn poll_result(&self) -> Result<Option<Value>, Value> {
-        let (lock, _) = &*self.inner;
-        let mut state = lock.lock().unwrap();
-        if let Some(val) = state.queue.pop_front() {
-            Self::finish_if_drained(&mut state);
-            return Ok(Some(val));
-        }
-        if state.drained_closed
-            && let Some(err) = state.failure.clone()
-        {
-            return Err(err);
-        }
-        Ok(None)
-    }
-
-    pub(crate) fn close(&self) {
-        let (lock, cvar) = &*self.inner;
-        let mut state = lock.lock().unwrap();
-        state.send_closed = true;
-        Self::finish_if_drained(&mut state);
-        cvar.notify_all();
-    }
-
-    pub(crate) fn fail(&self, error: Value) {
-        let (lock, cvar) = &*self.inner;
-        let mut state = lock.lock().unwrap();
-        state.send_closed = true;
-        state.failure = Some(error);
-        Self::finish_if_drained(&mut state);
-        cvar.notify_all();
-    }
-
-    pub(crate) fn closed_promise(&self) -> SharedPromise {
-        let (lock, _) = &*self.inner;
-        lock.lock().unwrap().closed_promise.clone()
-    }
-
-    pub(crate) fn can_send(&self) -> bool {
-        let (lock, _) = &*self.inner;
-        !lock.lock().unwrap().send_closed
-    }
-
-    pub(crate) fn add_supplier(&self, supplier_id: u64) {
-        let (lock, _) = &*self.inner;
-        lock.lock().unwrap().supplier_ids.push(supplier_id);
-    }
-
-    pub(crate) fn supplier_ids(&self) -> Vec<u64> {
-        let (lock, _) = &*self.inner;
-        lock.lock().unwrap().supplier_ids.clone()
-    }
-}
-
-impl PartialEq for SharedChannel {
-    fn eq(&self, other: &Self) -> bool {
-        Arc::ptr_eq(&self.inner, &other.inner)
-    }
-}
-
-impl PartialEq for Value {
-    fn eq(&self, other: &Self) -> bool {
-        let match_equals_pair_array = |attrs: &Arc<InstanceAttrs>, arr: &Arc<ArrayData>| {
-            if arr.len() != 2 {
-                return false;
-            }
-            let map = attrs.as_map();
-            let (Some(from), Some(matched)) = (map.get("from"), map.get("str")) else {
-                return false;
-            };
-            arr[0] == *from && arr[1] == *matched
-        };
-        match (self, other) {
-            (Value::Int(a), Value::Int(b)) => a == b,
-            (Value::BigInt(a), Value::BigInt(b)) => a == b,
-            (Value::BigInt(a), Value::Int(b)) | (Value::Int(b), Value::BigInt(a)) => {
-                **a == NumBigInt::from(*b)
-            }
-            (Value::Num(a), Value::Num(b)) => (a.is_nan() && b.is_nan()) || a == b,
-            (Value::Int(a), Value::Num(b)) => (*a as f64) == *b,
-            (Value::Num(a), Value::Int(b)) => *a == (*b as f64),
-            (Value::Str(a), Value::Str(b)) => a == b,
-            (Value::Bool(a), Value::Bool(b)) => a == b,
-            (Value::Range(a1, b1), Value::Range(a2, b2)) => a1 == a2 && b1 == b2,
-            (Value::RangeExcl(a1, b1), Value::RangeExcl(a2, b2)) => a1 == a2 && b1 == b2,
-            (Value::RangeExclStart(a1, b1), Value::RangeExclStart(a2, b2)) => a1 == a2 && b1 == b2,
-            (Value::RangeExclBoth(a1, b1), Value::RangeExclBoth(a2, b2)) => a1 == a2 && b1 == b2,
-            (
-                Value::GenericRange {
-                    start: s1,
-                    end: e1,
-                    excl_start: es1,
-                    excl_end: ee1,
-                },
-                Value::GenericRange {
-                    start: s2,
-                    end: e2,
-                    excl_start: es2,
-                    excl_end: ee2,
-                },
-            ) => es1 == es2 && ee1 == ee2 && s1 == s2 && e1 == e2,
-            (Value::Array(a, ..), Value::Array(b, ..)) => a == b,
-            (Value::Seq(a), Value::Seq(b)) => a == b,
-            (Value::Slip(a), Value::Slip(b)) => a == b,
-            (Value::Array(a, ..), Value::Seq(b))
-            | (Value::Seq(b), Value::Array(a, ..))
-            | (Value::Array(a, ..), Value::Slip(b))
-            | (Value::Slip(b), Value::Array(a, ..)) => a.items == **b,
-            (Value::Seq(a), Value::Slip(b)) | (Value::Slip(b), Value::Seq(a)) => {
-                a.as_ref() == b.as_ref()
-            }
-            (Value::Hash(a), Value::Hash(b)) => a == b,
-            (Value::Rat(a1, b1), Value::Rat(a2, b2)) => {
-                if *b1 == 0 && *b2 == 0 && *a1 == 0 && *a2 == 0 {
-                    return false; // NaN != NaN
-                }
-                a1 == a2 && b1 == b2
-            }
-            (Value::Rat(n, d), Value::Int(i)) | (Value::Int(i), Value::Rat(n, d)) => {
-                *d != 0 && *n == *i * *d
-            }
-            (Value::Rat(n, d), Value::Num(f)) | (Value::Num(f), Value::Rat(n, d)) => {
-                if *d == 0 {
-                    return false;
-                }
-                (*n as f64 / *d as f64) == *f
-            }
-            (Value::BigRat(an, ad), Value::BigRat(bn, bd)) => an == bn && ad == bd,
-            (Value::BigRat(n, d), Value::Int(i)) | (Value::Int(i), Value::BigRat(n, d)) => {
-                let (n, d) = (n.as_ref(), d.as_ref());
-                !d.is_zero() && *n == NumBigInt::from(*i) * d
-            }
-            (Value::BigRat(n, d), Value::Rat(rn, rd))
-            | (Value::Rat(rn, rd), Value::BigRat(n, d)) => {
-                let (n, d) = (n.as_ref(), d.as_ref());
-                !d.is_zero()
-                    && *rd != 0
-                    && n.clone() * NumBigInt::from(*rd) == NumBigInt::from(*rn) * d
-            }
-            (Value::BigRat(n, d), Value::Num(f)) | (Value::Num(f), Value::BigRat(n, d)) => {
-                let (n, d) = (n.as_ref(), d.as_ref());
-                if d.is_zero() {
-                    return false;
-                }
-                (n.to_f64().unwrap_or(0.0) / d.to_f64().unwrap_or(1.0)) == *f
-            }
-            (Value::Complex(r1, i1), Value::Complex(r2, i2)) => {
-                let f_eq = |a: &f64, b: &f64| (a.is_nan() && b.is_nan()) || a == b;
-                f_eq(r1, r2) && f_eq(i1, i2)
-            }
-            (Value::Complex(r, i), Value::Int(n)) | (Value::Int(n), Value::Complex(r, i)) => {
-                *i == 0.0 && *r == *n as f64
-            }
-            (Value::Complex(r, i), Value::Num(f)) | (Value::Num(f), Value::Complex(r, i)) => {
-                *i == 0.0 && *r == *f
-            }
-            (Value::Complex(r, i), Value::Rat(n, d)) | (Value::Rat(n, d), Value::Complex(r, i)) => {
-                *d != 0 && *i == 0.0 && *r == (*n as f64 / *d as f64)
-            }
-            (Value::Complex(r, i), Value::FatRat(n, d))
-            | (Value::FatRat(n, d), Value::Complex(r, i)) => {
-                *d != 0 && *i == 0.0 && *r == (*n as f64 / *d as f64)
-            }
-            (Value::Complex(r, i), Value::BigRat(n, d))
-            | (Value::BigRat(n, d), Value::Complex(r, i)) => {
-                let (n, d) = (n.as_ref(), d.as_ref());
-                !d.is_zero()
-                    && *i == 0.0
-                    && *r == (n.to_f64().unwrap_or(0.0) / d.to_f64().unwrap_or(1.0))
-            }
-            (Value::FatRat(a1, b1), Value::FatRat(a2, b2)) => a1 == a2 && b1 == b2,
-            (Value::Set(a, _), Value::Set(b, _)) => a == b,
-            (Value::Bag(a, _), Value::Bag(b, _)) => a == b,
-            (Value::Mix(a, _), Value::Mix(b, _)) => a == b,
-            (
-                Value::CompUnitDepSpec { short_name: a },
-                Value::CompUnitDepSpec { short_name: b },
-            ) => a == b,
-            (Value::Package(a), Value::Package(b)) => a == b,
-            (Value::Pair(ak, av), Value::Pair(bk, bv)) => ak == bk && av == bv,
-            (Value::ValuePair(ak, av), Value::ValuePair(bk, bv)) => ak == bk && av == bv,
-            (Value::Pair(ak, av), Value::ValuePair(bk, bv)) => {
-                matches!(bk.as_ref(), Value::Str(s) if s.as_str() == ak) && av == bv
-            }
-            (Value::ValuePair(ak, av), Value::Pair(bk, bv)) => {
-                matches!(ak.as_ref(), Value::Str(s) if s.as_str() == bk) && av == bv
-            }
-            (
-                Value::Enum {
-                    enum_type: at,
-                    key: ak,
-                    ..
-                },
-                Value::Enum {
-                    enum_type: bt,
-                    key: bk,
-                    ..
-                },
-            ) => at == bt && ak == bk,
-            (Value::Regex(a), Value::Regex(b)) => a == b,
-            (Value::RegexWithAdverbs(a), Value::RegexWithAdverbs(b)) => {
-                a.pattern == b.pattern
-                    && a.global == b.global
-                    && a.exhaustive == b.exhaustive
-                    && a.overlap == b.overlap
-                    && a.repeat == b.repeat
-                    && a.nth == b.nth
-                    && a.perl5 == b.perl5
-                    && a.pos == b.pos
-                    && a.continue_ == b.continue_
-                    && a.ignore_case == b.ignore_case
-                    && a.sigspace == b.sigspace
-                    && a.samecase == b.samecase
-                    && a.samespace == b.samespace
-            }
-            (
-                Value::Routine {
-                    package: ap,
-                    name: an,
-                    ..
-                },
-                Value::Routine {
-                    package: bp,
-                    name: bn,
-                    ..
-                },
-            ) => ap == bp && an == bn,
-            (
-                Value::Instance {
-                    class_name: a,
-                    attributes: aa,
-                    ..
-                },
-                Value::Instance {
-                    class_name: b,
-                    attributes: ba,
-                    ..
-                },
-            ) => a == b && aa == ba,
-            (
-                Value::Instance {
-                    class_name,
-                    attributes,
-                    ..
-                },
-                Value::Array(items, ..),
-            ) if class_name == "Match" => match_equals_pair_array(attributes, items),
-            (
-                Value::Array(items, ..),
-                Value::Instance {
-                    class_name,
-                    attributes,
-                    ..
-                },
-            ) if class_name == "Match" => match_equals_pair_array(attributes, items),
-            (
-                Value::Junction {
-                    kind: ak,
-                    values: av,
-                },
-                Value::Junction {
-                    kind: bk,
-                    values: bv,
-                },
-            ) => ak == bk && av == bv,
-            (Value::Sub(a), Value::Sub(b)) => a.id == b.id,
-            (Value::LazyList(a), Value::LazyList(b)) => Arc::ptr_eq(a, b),
-            (
-                Value::Version {
-                    parts: ap,
-                    plus: aplus,
-                    minus: aminus,
-                },
-                Value::Version {
-                    parts: bp,
-                    plus: bplus,
-                    minus: bminus,
-                },
-            ) => {
-                if aplus != bplus || aminus != bminus {
-                    return false;
-                }
-                // Normalize trailing zeroes: compare after stripping trailing Num(0)
-                let a_norm = Self::version_strip_trailing_zeros(ap);
-                let b_norm = Self::version_strip_trailing_zeros(bp);
-                a_norm == b_norm
-            }
-            (Value::Promise(a), Value::Promise(b)) => a == b,
-            (Value::Channel(a), Value::Channel(b)) => a == b,
-            (Value::Nil, Value::Nil) => true,
-            (
-                Value::Capture {
-                    positional: ap,
-                    named: an,
-                },
-                Value::Capture {
-                    positional: bp,
-                    named: bn,
-                },
-            ) => ap == bp && an == bn,
-            // Mixin (allomorphic types): compare inner values and mixin maps
-            (Value::Mixin(a_inner, a_mix), Value::Mixin(b_inner, b_mix)) => {
-                a_inner == b_inner && a_mix == b_mix
-            }
-            // Mixin vs non-Mixin: delegate to the inner value
-            (Value::Mixin(inner, _), other) | (other, Value::Mixin(inner, _)) => {
-                inner.as_ref() == other
-            }
-            // LazyThunk: compare cached values if available. Check Arc pointer
-            // identity first — the same thunk compared to itself must short-circuit
-            // before locking, since `a.cache` and `b.cache` are the *same* mutex
-            // and a non-reentrant std `Mutex` would deadlock on the second lock.
-            // (Hit by the END-phaser overlay's `v != orig_v` on a lexical bound to
-            // a `lazy { … }` thunk: captured and live env hold the same Arc.)
-            (Value::LazyThunk(a), Value::LazyThunk(b)) => {
-                if Arc::ptr_eq(a, b) {
-                    return true;
-                }
-                let a_cache = a.cache.lock().unwrap();
-                let b_cache = b.cache.lock().unwrap();
-                match (a_cache.as_ref(), b_cache.as_ref()) {
-                    (Some(av), Some(bv)) => av == bv,
-                    _ => false,
-                }
-            }
-            (Value::LazyThunk(thunk), other) | (other, Value::LazyThunk(thunk)) => {
-                let cache = thunk.cache.lock().unwrap();
-                if let Some(ref cached) = *cache {
-                    cached == other
-                } else {
-                    false
-                }
-            }
-            (Value::Uni(a), Value::Uni(b)) => a.form == b.form && a.text == b.text,
-            // ContainerRef: deref and compare inner values.
-            // Check Arc pointer identity first to avoid deadlock on same-Arc comparison.
-            (Value::ContainerRef(a), Value::ContainerRef(b)) => {
-                if Arc::ptr_eq(a, b) {
-                    return true;
-                }
-                let a_val = a.lock().unwrap().clone();
-                let b_val = b.lock().unwrap().clone();
-                a_val == b_val
-            }
-            (Value::ContainerRef(a), other) | (other, Value::ContainerRef(a)) => {
-                let a_val = a.lock().unwrap().clone();
-                a_val == *other
-            }
-            _ => false,
-        }
-    }
-}
-
-impl Value {
-    /// Create a decontainerized Proxy value (result of .VAR on a Proxy).
-    /// This Proxy won't be auto-FETCHed by method dispatch.
-    pub(crate) fn proxy_var_object(proxy: Value, _target_var: String) -> Self {
-        match proxy {
-            Value::Proxy {
-                fetcher,
-                storer,
-                subclass,
-                ..
-            } => Value::Proxy {
-                fetcher,
-                storer,
-                subclass,
-                decontainerized: true,
-            },
-            other => other,
-        }
-    }
-
-    // ---- Arc-wrapping convenience constructors ----
-    pub fn bigint(n: NumBigInt) -> Self {
-        Value::BigInt(Arc::new(n))
-    }
-    pub fn str(s: String) -> Self {
-        Value::Str(Arc::new(s))
-    }
-    pub fn str_from(s: &str) -> Self {
-        Value::Str(Arc::new(s.to_string()))
-    }
-    pub fn regex(s: String) -> Self {
-        Value::Regex(Arc::new(s))
-    }
-    pub fn mixin(inner: Value, overrides: HashMap<String, Value>) -> Self {
-        Value::Mixin(Arc::new(inner), Arc::new(overrides))
-    }
-    pub fn generic_range(start: Value, end: Value, excl_start: bool, excl_end: bool) -> Self {
-        Value::GenericRange {
-            start: Arc::new(start),
-            end: Arc::new(end),
-            excl_start,
-            excl_end,
-        }
-    }
-    pub fn array(items: Vec<Value>) -> Self {
-        Value::Array(Arc::new(ArrayData::new(items)), ArrayKind::List)
-    }
-    /// Create a Capture value. Boxes the positional/named payloads (the variant
-    /// stores them behind `Box` to keep `Value` small).
-    pub fn capture(positional: Vec<Value>, named: HashMap<String, Value>) -> Self {
-        Value::Capture {
-            positional: Box::new(positional),
-            named: Box::new(named),
-        }
-    }
-    /// Create a BigRat value. The numerator/denominator are boxed (the variant
-    /// stores them behind `Box` to keep `Value` small).
-    pub fn bigrat(num: NumBigInt, den: NumBigInt) -> Self {
-        Value::BigRat(Box::new(num), Box::new(den))
-    }
-    /// The HOW (meta-object) of a `CustomType` or `CustomTypeInstance`, if this
-    /// is one. Centralizes access now that both are boxed payloads.
-    pub fn custom_how(&self) -> Option<&Value> {
-        match self {
-            Value::CustomType(d) => Some(&d.how),
-            Value::CustomTypeInstance(d) => Some(&d.how),
-            _ => None,
-        }
-    }
-    /// The REPR name of a `CustomType` or `CustomTypeInstance`, if this is one.
-    pub fn custom_repr(&self) -> Option<&str> {
-        match self {
-            Value::CustomType(d) => Some(&d.repr),
-            Value::CustomTypeInstance(d) => Some(&d.repr),
-            _ => None,
-        }
-    }
-    /// Create a CustomType value (boxed payload).
-    pub fn custom_type(how: Box<Value>, repr: String, name: Symbol, id: u64) -> Self {
-        Value::CustomType(Box::new(CustomTypeData {
-            how,
-            repr,
-            name,
-            id,
-        }))
-    }
-    /// Create a Uni value (boxed payload).
-    pub fn uni(form: String, text: String) -> Self {
-        Value::Uni(Box::new(UniData { form, text }))
-    }
-    /// Create a CustomTypeInstance value (boxed payload).
-    pub fn custom_type_instance(
-        type_id: u64,
-        how: Box<Value>,
-        repr: String,
-        type_name: Symbol,
-        attributes: Arc<HashMap<String, Value>>,
-        id: u64,
-    ) -> Self {
-        Value::CustomTypeInstance(Box::new(CustomTypeInstanceData {
-            type_id,
-            how,
-            repr,
-            type_name,
-            attributes,
-            id,
-        }))
-    }
-    /// Create a true Array value (from [...] literals).
-    pub fn real_array(items: Vec<Value>) -> Self {
-        Value::Array(Arc::new(ArrayData::new(items)), ArrayKind::Array)
-    }
-    /// Create a shaped (multidimensional) Array value.
-    pub fn shaped_array(items: Vec<Value>) -> Self {
-        Value::Array(Arc::new(ArrayData::new(items)), ArrayKind::Shaped)
-    }
-    /// Build an `Arc<ArrayData>` from a plain element vector.
-    pub fn array_arc(items: Vec<Value>) -> Arc<ArrayData> {
-        Arc::new(ArrayData::new(items))
-    }
-    /// Rebuild an array's backing data with new elements, preserving the
-    /// embedded container type metadata of `like` (used by mutators that
-    /// reconstruct the vector, so a typed `Array[Int]` stays typed).
-    pub fn array_data_like(like: &ArrayData, items: Vec<Value>) -> Arc<ArrayData> {
-        Arc::new(ArrayData {
-            items,
-            value_type: like.value_type.clone(),
-            key_type: like.key_type.clone(),
-            declared_type: like.declared_type.clone(),
-            default: like.default.clone(),
-            shape: like.shape.clone(),
-        })
-    }
-    /// Construct a `Value::Hash`. Accepts either a bare `HashMap` (fresh hash)
-    /// or a `HashData` (a cloned/rebuilt hash whose container metadata is then
-    /// preserved) via `Into<HashData>`.
-    pub fn hash(map: impl Into<HashData>) -> Self {
-        Value::Hash(Arc::new(map.into()))
-    }
-
-    /// Build an `Arc<HashData>` from a map or `HashData`. Lets call sites that
-    /// constructed `Value::Hash(Arc::new(x))` keep their shape as
-    /// `Value::Hash(Value::hash_arc(x))` while the variant moved to `HashData`.
-    pub fn hash_arc(map: impl Into<HashData>) -> Arc<HashData> {
-        Arc::new(map.into())
-    }
-
-    /// Coerce a value into item context (`.item` method).
-    /// Arrays get their kind itemized, hashes get their itemization flag set
-    /// (mirroring `ArrayKind` — the value stays a `Value::Hash` so it never
-    /// leaks a wrapper to value operations), other values get wrapped in Scalar.
-    pub fn item(self) -> Self {
-        match self {
-            Value::Array(items, kind) => Value::Array(items, kind.itemize()),
-            Value::Hash(h) => Value::Hash(Self::hash_arc_itemized(h)),
-            other => other,
-        }
-    }
-
-    /// Return a `Value::Hash` Arc with its itemization flag set (copy-on-write
-    /// only when the flag actually changes, so `.WHICH` identity is preserved
-    /// for an already-itemized hash). Used by the `Itemize`/`ItemizeVar` opcodes
-    /// and `.item` to mark a `$`-sourced hash as a single list-context element.
-    pub(crate) fn hash_arc_itemized(h: Arc<HashData>) -> Arc<HashData> {
-        if h.itemized {
-            h
-        } else {
-            let mut data = (*h).clone();
-            data.itemized = true;
-            Arc::new(data)
-        }
-    }
-
-    pub(crate) fn hash_arc_deitemized(h: Arc<HashData>) -> Arc<HashData> {
-        if h.itemized {
-            let mut data = (*h).clone();
-            data.itemized = false;
-            Arc::new(data)
-        } else {
-            h
-        }
-    }
-
-    /// Read through a `ContainerRef` and apply `f` to the inner value WITHOUT
-    /// cloning it. Non-ContainerRef values are passed to `f` as-is. This is the
-    /// canonical non-cloning ContainerRef-read chokepoint (the ContainerRef axis of
-    /// the decont family); prefer it over hand-rolled `arc.lock().unwrap()` reads.
-    pub fn with_deref<R>(&self, f: impl FnOnce(&Value) -> R) -> R {
-        match self {
-            Value::ContainerRef(arc) => f(&arc.lock().unwrap()),
-            other => f(other),
-        }
-    }
-
-    /// Read through a `ContainerRef`, returning an owned clone of the inner value.
-    /// Non-ContainerRef values are cloned as-is. Use [`Value::with_deref`] instead
-    /// when you only need to read the inner value (it avoids the clone).
-    pub fn deref_container(&self) -> Value {
-        self.with_deref(Value::clone)
-    }
-
-    /// Owned counterpart of [`Value::deref_container`]: read through a
-    /// `ContainerRef` BY VALUE, cloning ONLY the inner value when `self` is a
-    /// ContainerRef; non-ContainerRef values move through with no clone. This is
-    /// the canonical move-friendly read chokepoint for hot read opcodes
-    /// (`GetLocal`/`GetGlobal`), mirroring how [`Value::into_descalarized`] is the
-    /// owned variant of [`Value::descalarize`]. Single-level only (a nested
-    /// `ContainerRef` is unwrapped one cell), and it does NOT force an inner
-    /// `LazyThunk` nor strip an inner `Scalar` — matching the prior hand-rolled
-    /// `arc.lock().unwrap().clone()` reads it replaces.
-    pub fn into_deref(self) -> Value {
-        match self {
-            Value::ContainerRef(arc) => arc.lock().unwrap().clone(),
-            other => other,
-        }
-    }
-
-    /// Assign a value into a `ContainerRef`.
-    /// Returns `true` if the value was a ContainerRef and the assignment happened.
-    pub fn assign_into_container(&self, new_val: Value) -> bool {
-        if let Value::ContainerRef(arc) = self {
-            let mut inner = arc.lock().unwrap();
-            *inner = new_val;
-            true
-        } else {
-            false
-        }
-    }
-
-    /// Create a new shared container holding this value.
-    pub fn into_container_ref(self) -> Value {
-        Value::ContainerRef(Arc::new(Mutex::new(self)))
-    }
-
-    /// Assign `val` into an array/hash element `slot`. When the slot already
-    /// holds a `ContainerRef` cell (a Phase 2 `:=`-bound element), write
-    /// *through* the cell so every alias of that element observes the new
-    /// value; otherwise replace the slot in place. This is the single element
-    /// write chokepoint that keeps a bound element's alias live across writes.
-    pub fn assign_element_slot(slot: &mut Value, val: Value) {
-        if let Value::ContainerRef(cell) = slot {
-            *cell.lock().unwrap() = val;
-        } else {
-            *slot = val;
-        }
-    }
-
-    /// Hash element write chokepoint (Phase 2 Stage 0). The hash analogue of
-    /// [`assign_element_slot`]: if the existing entry at `key` is a
-    /// `ContainerRef` cell, write *through* it (preserving any `:=` binding);
-    /// otherwise insert or replace the entry as a bare value.
-    ///
-    /// This is behavior-invariant until hash elements are promoted to cells
-    /// (Phase 2 Stage 1), because no hash currently stores `ContainerRef`
-    /// entries, so every call collapses to a plain insert/replace. Routing all
-    /// hash-element writes through this single chokepoint is the prerequisite
-    /// for that promotion: a naive promotion without it broke array-through-hash
-    /// traversal (nested.t 30->7), see `docs/container-identity.md`.
-    pub fn hash_insert_through(map: &mut HashMap<String, Value>, key: String, val: Value) {
-        match map.get_mut(&key) {
-            Some(slot) => Value::assign_element_slot(slot, val),
-            None => {
-                map.insert(key, val);
-            }
-        }
-    }
-
-    pub fn is_container_ref(&self) -> bool {
-        matches!(self, Value::ContainerRef(_))
-    }
-
-    /// Autovivify a hash entry: if the key doesn't exist, insert an empty Hash.
-    /// Returns a `HashEntryRef` pointing to the entry in the parent hash.
-    /// Uses interior mutation of the `Arc<HashMap>` so that **all** clones of
-    /// the same `Arc` observe the change.  This relies on no `&HashMap` borrow
-    /// being live across the call (the aliased-mutation contract in
-    /// `aliased_mut.rs`); cross-thread sharing of the same `Arc` is the tracked,
-    /// pre-existing gap documented there.
-    ///
-    /// Look up a key in a Hash value by string key.
-    pub fn hash_get_str(&self, key: &str) -> Option<Value> {
-        match self {
-            Value::Hash(arc) => arc.get(key).cloned(),
-            Value::Mixin(inner, _) => inner.hash_get_str(key),
-            Value::Scalar(inner) => inner.hash_get_str(key),
-            _ => None,
-        }
-    }
-
-    /// Returns `None` if `self` is not a `Value::Hash`.
-    pub fn hash_autovivify(&self, key: &str) -> Option<Value> {
-        if let Value::Hash(arc) = self {
-            // SAFETY: aliased in-place mutation of a shared container; see
-            // `arc_contents_mut`. No borrow into the map is live across the write.
-            let data = unsafe { arc_contents_mut(arc) };
-            if !data.map.contains_key(key) {
-                let new_hash = Value::hash(HashMap::new());
-                data.map.insert(key.to_string(), new_hash);
-            }
-            Some(Value::HashEntryRef {
-                hash: arc.clone(),
-                path: vec![key.to_string()],
-            })
-        } else {
-            None
-        }
-    }
-
-    /// Autovivifying hash element access for bind descent — the hash analogue of
-    /// [`array_slot_ref`] (Phase 2). Instead of the stale `HashEntryRef`
-    /// back-reference, it returns a first-class value that survives COW:
-    /// - an existing `ContainerRef` cell is returned as-is (already aliased);
-    /// - an existing container leaf (Array/Hash) is returned by value — it shares
-    ///   the inner Arc, so descent and the eventual leaf write land in the same
-    ///   physical container (no back-reference needed, like the lazy op);
-    /// - an existing scalar leaf is promoted in place to a shared `ContainerRef`
-    ///   cell and that cell is returned (the bind aliases it by cell identity);
-    /// - a missing key is autovivified to an empty Hash (the old descent
-    ///   behavior) and returned by value (shared Arc).
-    pub fn hash_autovivify_cell(&self, key: &str) -> Option<Value> {
-        if let Value::Hash(arc) = self {
-            // SAFETY: aliased in-place mutation of a shared container; see
-            // `arc_contents_mut`. No borrow into the map is live across the write.
-            let data = unsafe { arc_contents_mut(arc) };
-            match data.map.get_mut(key) {
-                Some(Value::ContainerRef(cell)) => Some(Value::ContainerRef(cell.clone())),
-                Some(elem @ (Value::Array(..) | Value::Hash(..))) => Some(elem.clone()),
-                Some(elem) => {
-                    let cell = Arc::new(Mutex::new(std::mem::replace(elem, Value::Nil)));
-                    *elem = Value::ContainerRef(cell.clone());
-                    Some(Value::ContainerRef(cell))
-                }
-                None => {
-                    let new_hash = Value::hash(HashMap::new());
-                    data.map.insert(key.to_string(), new_hash.clone());
-                    Some(new_hash)
-                }
-            }
-        } else {
-            None
-        }
-    }
-
-    /// Bind to hash element `key`, promoting it to a first-class container
-    /// (Phase 2 Stage 1) — the hash analogue of [`array_slot_ref`]. An existing
-    /// *scalar* leaf is replaced in place with a shared `ContainerRef` cell
-    /// (reusing one if already present), and that same cell is returned so the
-    /// binding aliases the element by **cell identity**, surviving COW clones of
-    /// any enclosing container on a later write (the staleness that the old
-    /// `HashEntryRef` back-reference suffers for deep `%h<a><b>` paths).
-    ///
-    /// An existing *container* leaf (Array/Hash) is an intermediate level of a
-    /// deeper path (`%h<a><b>`, `%h<a>[1]`); it keeps the old `HashEntryRef` so
-    /// the deeper traversal resolves through the shared inner Arc and the
-    /// eventual leaf promotion lands in the physical map the entry points to.
-    /// A *missing* key stays lazy (no entry created) — promotion is deferred to
-    /// the first write (a `HashEntryRef` token carries the path until then).
-    ///
-    /// Reads decontainerize at the single chokepoint (`resolve_hash_entry`);
-    /// writes go through `hash_insert_through` (Stage 0).
-    pub fn hash_slot_ref(&self, key: &str, terminal: bool) -> Option<Value> {
-        if let Value::Hash(arc) = self {
-            // SAFETY: aliased in-place mutation of a shared container; see
-            // `arc_contents_mut`. No borrow into the map is live across the write.
-            let data = unsafe { arc_contents_mut(arc) };
-            match data.map.get_mut(key) {
-                Some(Value::ContainerRef(cell)) => Some(Value::ContainerRef(cell.clone())),
-                Some(elem @ (Value::Array(..) | Value::Hash(..))) if !terminal => {
-                    // Intermediate container: return the element value
-                    // itself — it shares the inner Arc, so the eventual
-                    // leaf promotion by the next index op lands in the
-                    // physical map the entry points to (Stage 2: no
-                    // `HashEntryRef` back-reference needed).
-                    Some(elem.clone())
-                }
-                Some(elem) => {
-                    let cell = Arc::new(Mutex::new(std::mem::replace(elem, Value::Nil)));
-                    *elem = Value::ContainerRef(cell.clone());
-                    Some(Value::ContainerRef(cell))
-                }
-                None => Some(Value::HashEntryRef {
-                    hash: arc.clone(),
-                    path: vec![key.to_string()],
-                }),
-            }
-        } else {
-            None
-        }
-    }
-
-    /// Autovivify a hash entry with a scalar value (for binding/assignment).
-    /// Inserts the given value at the key if missing, or replaces the existing value.
-    /// Returns the value stored at the key after the operation.
-    /// Uses the same interior-mutation approach as `hash_autovivify`.
-    pub fn hash_assign_at(&self, key: &str, val: Value) -> Option<Value> {
-        if let Value::Hash(arc) = self {
-            // SAFETY: aliased in-place mutation of a shared container; see
-            // `arc_contents_mut`. No borrow into the map is live across the write.
-            let data = unsafe { arc_contents_mut(arc) };
-            Value::hash_insert_through(&mut data.map, key.to_string(), val.clone());
-            Some(val)
-        } else {
-            None
-        }
-    }
-
-    /// Read the current value a `HashEntryRef` points to, walking its `path`
-    /// READ-ONLY (no autovivification). Returns `Any` if any intermediate level
-    /// is missing or not a hash, or if the terminal key is absent — so a bind to
-    /// a not-yet-existent entry reads as `Any` without polluting `:exists`.
-    pub fn hash_entry_read(&self) -> Value {
-        let Value::HashEntryRef { hash, path } = self else {
-            return self.clone();
-        };
-        let any = || Value::Package(crate::symbol::Symbol::intern("Any"));
-        let mut cur = hash.clone();
-        for k in &path[..path.len() - 1] {
-            let ptr = Arc::as_ptr(&cur);
-            match unsafe { (*ptr).get(k.as_str()) } {
-                Some(Value::Hash(inner)) => cur = inner.clone(),
-                _ => return any(),
-            }
-        }
-        let last = path.last().unwrap();
-        let ptr = Arc::as_ptr(&cur);
-        unsafe { (*ptr).get(last.as_str()).cloned().unwrap_or_else(any) }
-    }
-
-    /// Walk-CREATE the intermediate hashes of a `HashEntryRef`'s `path` and
-    /// return the `(terminal hash Arc, terminal key)` so the caller can insert.
-    /// Missing/non-hash intermediate levels are replaced with fresh empty hashes
-    /// (interior mutation, so all holders of the shared Arc observe them).
-    pub fn hash_entry_terminal(&self) -> Option<(Arc<HashData>, String)> {
-        let Value::HashEntryRef { hash, path } = self else {
-            return None;
-        };
-        let mut cur = hash.clone();
-        for k in &path[..path.len() - 1] {
-            // SAFETY: aliased in-place mutation of a shared hash; see
-            // `arc_contents_mut`. No borrow into the map is live across the write.
-            let next = {
-                let data = unsafe { arc_contents_mut(&cur) };
-                match data.map.get(k) {
-                    Some(Value::Hash(inner)) => inner.clone(),
-                    _ => {
-                        let new_hash = Value::hash(HashMap::new());
-                        let arc = match &new_hash {
-                            Value::Hash(a) => a.clone(),
-                            _ => unreachable!(),
-                        };
-                        Value::hash_insert_through(&mut data.map, k.clone(), new_hash);
-                        arc
-                    }
-                }
-            };
-            cur = next;
-        }
-        Some((cur, path.last().unwrap().clone()))
-    }
-
-    /// Write a value through a `HashEntryRef`, walk-creating intermediate hashes
-    /// and inserting at the terminal key (interior mutation).
-    pub fn hash_entry_write(&self, val: Value) {
-        if let Some((arc, key)) = self.hash_entry_terminal() {
-            // SAFETY: aliased in-place mutation of a shared hash; see
-            // `arc_contents_mut`. No borrow into the map is live across the write.
-            let data = unsafe { arc_contents_mut(&arc) };
-            Value::hash_insert_through(&mut data.map, key, val);
-        }
-    }
-
-    /// Push a value to an Array in-place using interior mutation.
-    /// This allows shared references (Arc refcount > 1) to see the mutation,
-    /// matching Raku's container semantics where all references share state.
-    /// Safety: same assumptions as hash_autovivify — callers ensure no
-    /// concurrent reads/writes to the same Arc.
-    pub fn array_push_in_place(&self, val: Value) -> bool {
-        if let Value::Array(arc, _) = self {
-            // SAFETY: aliased in-place mutation of a shared container; see
-            // `arc_contents_mut`. No borrow into the items is live across the push.
-            let data = unsafe { arc_contents_mut(arc) };
-            data.items.push(val);
-            true
-        } else {
-            false
-        }
-    }
-
-    /// Bind to array element `idx`, promoting it to a first-class container
-    /// (Phase 2). The element is replaced in place with a shared
-    /// `ContainerRef` cell (reusing an existing one), and that same cell is
-    /// returned so the binding aliases the element by **cell identity**. Unlike
-    /// the old array element back-reference (an array-Arc + index back-reference, which goes
-    /// stale when an enclosing container is COW-cloned on a later write), the
-    /// `Arc<Mutex>` cell is shared on every clone, so the alias survives
-    /// arbitrarily deep `$struct[..]<..>[..]` paths. Reads decontainerize the
-    /// element at the single read chokepoint (`resolve_array_entry`).
-    pub fn array_slot_ref(&self, idx: usize, terminal: bool) -> Option<Value> {
-        if let Value::Array(arc, _kind) = self {
-            // SAFETY: aliased in-place mutation of a shared container; see
-            // `arc_contents_mut`. No borrow into the items is live across the
-            // growth/promotion below.
-            let data = unsafe { arc_contents_mut(arc) };
-            while data.len() <= idx {
-                data.push(Value::Nil);
-            }
-            let elem = &mut data[idx];
-            if let Value::ContainerRef(cell) = elem {
-                return Some(Value::ContainerRef(cell.clone()));
-            }
-            // Only promote a *scalar* leaf to a cell. A container element
-            // (Array/Hash) is an intermediate level of a deeper path
-            // (`$s[1][1]`, `$s[1]<k>`); return the element value itself —
-            // it shares the inner Arc, so the eventual leaf promotion by
-            // the next index op lands in the same physical Vec/HashMap the
-            // stored element points to (Stage 2: no array element back-reference
-            // back-reference needed).
-            if !terminal && matches!(elem, Value::Array(..) | Value::Hash(..)) {
-                return Some(elem.clone());
-            }
-            let cell = Arc::new(Mutex::new(std::mem::replace(elem, Value::Nil)));
-            *elem = Value::ContainerRef(cell.clone());
-            Some(Value::ContainerRef(cell))
-        } else {
-            None
-        }
-    }
-
-    /// Encode a Value as a hash key string.
-    /// Regex values are encoded with a special prefix to preserve their identity.
-    pub fn hash_key_encode(val: &Value) -> String {
-        match val {
-            Value::Regex(pattern) => {
-                format!("\0rx:{}", pattern)
-            }
-            other => other.to_string_value(),
-        }
-    }
-
-    /// Decode a hash key string back to a Value.
-    /// Regex-encoded keys (with \0rx: prefix) are restored to Regex values.
-    pub fn hash_key_decode(key: &str) -> Value {
-        if let Some(pattern) = key.strip_prefix("\0rx:") {
-            return Value::regex(pattern.to_string());
-        }
-        Value::str(key.to_string())
-    }
-
-    // --- decont family (see docs/container-identity.md §3) ---
-    // mutsu has THREE "decontainerize" operations on THREE different axes.
-    // They are intentionally NOT fused into one helper:
-    //   - Value::descalarize / into_descalarized — strips `Scalar` ($(...)), RECURSIVE
-    //   - Value::with_deref / deref_container / into_deref — reads through `ContainerRef` (:=),
-    //     single cell. with_deref is non-cloning; deref_container clones from &self;
-    //     into_deref consumes self, cloning only the inner of a ContainerRef (move otherwise).
-    //   - ArrayKind::decontainerize               — strips `ItemList`/`ItemArray` flag (list flatten)
-    // A "full decont" that strips all three is deferred to Phase 1+; it must NOT be
-    // applied at lvalue/container-requiring sites (is-rw writeback, :=, .VAR, =:=,
-    // take-rw, autoviv slot-refs), which need the live cell or the container variant.
-
-    /// Unwrap a `Scalar` container, returning a reference to the inner value.
-    /// RECURSIVE: nested `$($(...))` are fully stripped. Non-Scalar values are
-    /// returned as-is. See the decont family note above; this is the Scalar axis only.
-    pub fn descalarize(&self) -> &Value {
-        match self {
-            Value::Scalar(inner) => inner.descalarize(),
-            other => other,
-        }
-    }
-
-    /// Owned, recursive `Scalar`-strip. Same axis/semantics as [`Value::descalarize`]
-    /// but consumes `self` and returns the inner value by value (no extra clone for
-    /// callers that already own the value). Canonical replacement for the former
-    /// `runtime::methods_mut::strip_scalar`.
-    pub fn into_descalarized(self) -> Value {
-        match self {
-            Value::Scalar(inner) => inner.into_descalarized(),
-            other => other,
-        }
-    }
-    pub fn set(s: HashSet<String>) -> Self {
-        Value::Set(Arc::new(SetData::new(s)), false)
-    }
-    pub fn set_hash(s: HashSet<String>) -> Self {
-        Value::Set(Arc::new(SetData::new(s)), true)
-    }
-    /// Create a Set with preserved original key types.
-    pub fn set_typed(elements: HashSet<String>, original_keys: HashMap<String, Value>) -> Self {
-        Value::Set(
-            Arc::new(SetData::with_original_keys(elements, original_keys)),
-            false,
-        )
-    }
-    /// Create a SetHash with preserved original key types.
-    pub fn set_hash_typed(
-        elements: HashSet<String>,
-        original_keys: HashMap<String, Value>,
-    ) -> Self {
-        Value::Set(
-            Arc::new(SetData::with_original_keys(elements, original_keys)),
-            true,
-        )
-    }
-    /// Convert an i64 count map to the arbitrary-precision BigInt map BagData uses.
-    fn bag_counts_from_i64(m: HashMap<String, i64>) -> HashMap<String, NumBigInt> {
-        m.into_iter()
-            .map(|(k, v)| (k, NumBigInt::from(v)))
-            .collect()
-    }
-    pub fn bag(m: HashMap<String, i64>) -> Self {
-        Value::Bag(Arc::new(BagData::new(Self::bag_counts_from_i64(m))), false)
-    }
-    pub fn bag_hash(m: HashMap<String, i64>) -> Self {
-        Value::Bag(Arc::new(BagData::new(Self::bag_counts_from_i64(m))), true)
-    }
-    /// Create a Bag from an arbitrary-precision BigInt count map.
-    pub fn bag_big(m: HashMap<String, NumBigInt>) -> Self {
-        Value::Bag(Arc::new(BagData::new(m)), false)
-    }
-    /// Create a BagHash from an arbitrary-precision BigInt count map.
-    pub fn bag_hash_big(m: HashMap<String, NumBigInt>) -> Self {
-        Value::Bag(Arc::new(BagData::new(m)), true)
-    }
-    /// Create a Bag with preserved original key types.
-    pub fn bag_typed(counts: HashMap<String, i64>, original_keys: HashMap<String, Value>) -> Self {
-        Value::Bag(
-            Arc::new(BagData::with_original_keys(
-                Self::bag_counts_from_i64(counts),
-                original_keys,
-            )),
-            false,
-        )
-    }
-    /// Create a Bag with preserved original key types from a BigInt count map.
-    pub fn bag_typed_big(
-        counts: HashMap<String, NumBigInt>,
-        original_keys: HashMap<String, Value>,
-    ) -> Self {
-        Value::Bag(
-            Arc::new(BagData::with_original_keys(counts, original_keys)),
-            false,
-        )
-    }
-    /// Create a BagHash with preserved original key types.
-    pub fn bag_hash_typed(
-        counts: HashMap<String, i64>,
-        original_keys: HashMap<String, Value>,
-    ) -> Self {
-        Value::Bag(
-            Arc::new(BagData::with_original_keys(
-                Self::bag_counts_from_i64(counts),
-                original_keys,
-            )),
-            true,
-        )
-    }
-    /// Create a BagHash with preserved original key types from a BigInt count map.
-    pub fn bag_hash_typed_big(
-        counts: HashMap<String, NumBigInt>,
-        original_keys: HashMap<String, Value>,
-    ) -> Self {
-        Value::Bag(
-            Arc::new(BagData::with_original_keys(counts, original_keys)),
-            true,
-        )
-    }
-    pub fn mix(mut m: HashMap<String, f64>) -> Self {
-        m.retain(|_, weight| *weight != 0.0);
-        Value::Mix(Arc::new(MixData::new(m)), false)
-    }
-    pub fn mix_hash(mut m: HashMap<String, f64>) -> Self {
-        m.retain(|_, weight| *weight != 0.0);
-        Value::Mix(Arc::new(MixData::new(m)), true)
-    }
-    pub fn mix_with_original_keys(
-        mut weights: HashMap<String, f64>,
-        original_keys: HashMap<String, Value>,
-    ) -> Self {
-        weights.retain(|_, weight| *weight != 0.0);
-        Value::Mix(
-            Arc::new(MixData::with_original_keys(weights, original_keys)),
-            false,
-        )
-    }
-    pub fn mix_hash_with_original_keys(
-        mut weights: HashMap<String, f64>,
-        original_keys: HashMap<String, Value>,
-    ) -> Self {
-        weights.retain(|_, weight| *weight != 0.0);
-        Value::Mix(
-            Arc::new(MixData::with_original_keys(weights, original_keys)),
-            true,
-        )
-    }
-    pub fn slip(items: Vec<Value>) -> Self {
-        Value::Slip(Arc::new(items))
-    }
-    pub fn junction(kind: JunctionKind, values: Vec<Value>) -> Self {
-        Value::Junction {
-            kind,
-            values: Arc::new(values),
-        }
-    }
-
-    /// Create a new Sub value wrapping the given SubData in an Arc.
-    pub(crate) fn make_sub(
-        package: Symbol,
-        name: Symbol,
-        params: Vec<String>,
-        param_defs: Vec<ParamDef>,
-        body: Vec<Stmt>,
-        is_rw: bool,
-        env: Env,
-    ) -> Self {
-        Value::Sub(Arc::new(SubData {
-            package,
-            name,
-            params,
-            param_defs,
-            body,
-            is_rw,
-            is_raw: false,
-            env,
-            assumed_positional: Vec::new(),
-            assumed_named: HashMap::new(),
-            id: next_instance_id(),
-            empty_sig: false,
-            is_bare_block: false,
-            compiled_code: None,
-            deprecated_message: None,
-            source_line: None,
-            source_file: None,
-            owned_captures: Vec::new(),
-            upvalues: Vec::new(),
-        }))
-    }
-
-    /// Create a new Sub value with an explicit id.
-    #[allow(clippy::too_many_arguments)]
-    pub(crate) fn make_sub_with_id(
-        package: Symbol,
-        name: Symbol,
-        params: Vec<String>,
-        param_defs: Vec<ParamDef>,
-        body: Vec<Stmt>,
-        is_rw: bool,
-        env: Env,
-        id: u64,
-    ) -> Self {
-        Value::Sub(Arc::new(SubData {
-            package,
-            name,
-            params,
-            param_defs,
-            body,
-            is_rw,
-            is_raw: false,
-            env,
-            assumed_positional: Vec::new(),
-            assumed_named: HashMap::new(),
-            id,
-            empty_sig: false,
-            is_bare_block: false,
-            compiled_code: None,
-            deprecated_message: None,
-            source_line: None,
-            source_file: None,
-            owned_captures: Vec::new(),
-            upvalues: Vec::new(),
-        }))
-    }
-
-    /// Access SubData fields if this is a Sub (or upgraded WeakSub).
-    #[allow(dead_code)]
-    pub(crate) fn as_sub(&self) -> Option<&SubData> {
-        match self {
-            Value::Sub(data) => Some(data),
-            _ => None,
-        }
-    }
-
-    /// Upgrade a WeakSub to a Sub, or return Nil if expired.
-    #[allow(dead_code)]
-    pub(crate) fn upgrade_weak(&self) -> Value {
-        match self {
-            Value::WeakSub(weak) => match weak.upgrade() {
-                Some(strong) => Value::Sub(strong),
-                None => Value::Nil,
-            },
-            other => other.clone(),
-        }
-    }
-
-    pub(crate) fn make_instance(class_name: Symbol, attributes: HashMap<String, Value>) -> Self {
-        let id = next_instance_id();
-        Self::make_instance_with_id(class_name, attributes, id)
-    }
-
-    pub(crate) fn make_instance_without_destroy(
-        class_name: Symbol,
-        attributes: HashMap<String, Value>,
-    ) -> Self {
-        Self::make_instance_with_destroy(class_name, attributes, false)
-    }
-
-    /// Build an instance with the given id and a fresh attribute cell. Used for
-    /// constructing a value with an explicit id (genuinely new instances, or
-    /// sentinel ids). Cross-frame sharing of mutations comes from cloning an
-    /// existing instance's `Arc<InstanceAttrs>` (see [`Value::instance_sharing_cell`]),
-    /// not from this constructor — so this never reuses another holder's cell.
-    pub(crate) fn make_instance_with_id(
-        class_name: Symbol,
-        attributes: HashMap<String, Value>,
-        id: u64,
-    ) -> Self {
-        Value::Instance {
-            class_name,
-            attributes: Arc::new(InstanceAttrs::new(class_name, attributes, id, true)),
-            id,
-        }
-    }
-
-    /// Phase 3 registry-removal: return a `Value::Instance` that SHARES `attrs`'s
-    /// live cell, optionally under a new `class_name` (rebless / role mixin). This
-    /// replaces the `make_instance_with_id` rebuild branch, which reused the cell
-    /// by looking it up in the global `instance_cells` registry. Sharing the
-    /// `Arc<InstanceAttrs>` directly keeps in-place mutations visible to every
-    /// existing alias and to the returned value, without the registry.
-    pub(crate) fn instance_sharing_cell(
-        attrs: &Arc<InstanceAttrs>,
-        class_name: Symbol,
-        id: u64,
-    ) -> Value {
-        debug_assert_eq!(attrs.id, id, "instance_sharing_cell id mismatch");
-        let attributes = if attrs.class_name == class_name {
-            Arc::clone(attrs)
-        } else {
-            Arc::new(attrs.with_class(class_name))
-        };
-        Value::Instance {
-            class_name,
-            attributes,
-            id,
-        }
-    }
-
-    /// Phase 3 registry-removal: write `map` into `attrs`'s shared cell in place
-    /// and return a `Value::Instance` aliasing that same cell. The single helper
-    /// for the common writeback-then-rebuild pattern: it replaces the paired
-    /// `overwrite_instance_bindings_by_identity(..) + make_instance_with_id(..)`.
-    pub(crate) fn write_back_sharing(
-        attrs: &Arc<InstanceAttrs>,
-        class_name: Symbol,
-        map: HashMap<String, Value>,
-        id: u64,
-    ) -> Value {
-        attrs.commit_attrs(map);
-        Value::instance_sharing_cell(attrs, class_name, id)
-    }
-
-    /// An independent snapshot for `temp`/`let` saves. An instance is deep-copied
-    /// into a fresh, *unregistered* cell, so a later in-place mutation through
-    /// the live shared cell does not alter the saved value (pre-Stage-1 this
-    /// independence came for free from copy-on-write). Arrays/hashes keep their
-    /// CoW `Arc` (forked on first mutation), so a plain clone is already
-    /// independent for them.
-    pub(crate) fn into_temp_snapshot(self) -> Value {
-        match self {
-            Value::Instance {
-                class_name,
-                attributes,
-                id,
-            } => Value::Instance {
-                class_name,
-                attributes: Arc::new((*attributes).clone()),
-                id,
-            },
-            other => other,
-        }
-    }
-
-    fn make_instance_with_destroy(
-        class_name: Symbol,
-        attributes: HashMap<String, Value>,
-        queue_destroy: bool,
-    ) -> Self {
-        let id = next_instance_id();
-        Value::Instance {
-            class_name,
-            attributes: Arc::new(InstanceAttrs::new(
-                class_name,
-                attributes,
-                id,
-                queue_destroy,
-            )),
-            id,
-        }
-    }
-
-    /// Create an Instant value from the current system time.
-    pub(crate) fn make_instant_now() -> Self {
-        let posix = current_time_secs_f64();
-        let tai = crate::builtins::methods_0arg::temporal::posix_to_instant(posix);
-        let mut attrs = HashMap::new();
-        attrs.insert("value".to_string(), Value::Num(tai));
-        Value::make_instance(Symbol::intern("Instant"), attrs)
-    }
-
-    /// Create a Match object with positional captures.
-    pub(crate) fn make_match_object_with_captures(
-        matched: String,
-        from: i64,
-        to: i64,
-        positional: &[String],
-        named: &HashMap<String, Vec<String>>,
-    ) -> Self {
-        Self::make_match_object_full(
-            matched,
-            from,
-            to,
-            positional,
-            named,
-            &HashMap::new(),
-            &[],
-            &[],
-            &[],
-            None,
-        )
-    }
-
-    /// Create a Match object with positional captures and original string.
-    #[allow(clippy::too_many_arguments)]
-    pub(crate) fn make_match_object_full(
-        matched: String,
-        from: i64,
-        to: i64,
-        positional: &[String],
-        named: &HashMap<String, Vec<String>>,
-        named_subcaps: &HashMap<String, Vec<crate::runtime::RegexCaptures>>,
-        positional_subcaps: &[Option<crate::runtime::RegexCaptures>],
-        positional_quantified: &[Option<Vec<crate::runtime::QuantifiedCaptureEntry>>],
-        positional_nil: &[bool],
-        orig: Option<&str>,
-    ) -> Self {
-        Self::make_match_object_full_q(
-            matched,
-            from,
-            to,
-            positional,
-            named,
-            named_subcaps,
-            positional_subcaps,
-            positional_quantified,
-            positional_nil,
-            orig,
-            &HashSet::new(),
-        )
-    }
-
-    /// Like make_match_object_full but with named_quantified tracking.
-    #[allow(clippy::too_many_arguments)]
-    pub(crate) fn make_match_object_full_q(
-        matched: String,
-        from: i64,
-        to: i64,
-        positional: &[String],
-        named: &HashMap<String, Vec<String>>,
-        named_subcaps: &HashMap<String, Vec<crate::runtime::RegexCaptures>>,
-        positional_subcaps: &[Option<crate::runtime::RegexCaptures>],
-        positional_quantified: &[Option<Vec<crate::runtime::QuantifiedCaptureEntry>>],
-        positional_nil: &[bool],
-        orig: Option<&str>,
-        named_quantified: &HashSet<String>,
-    ) -> Self {
-        fn make_capture_match(s: &str, orig: Option<&str>, search_from: usize) -> Value {
-            let mut attrs = HashMap::new();
-            attrs.insert("str".to_string(), Value::str(s.to_string()));
-            // Try to find the captured text's position within the original string
-            let (cap_from, cap_to) = if let Some(o) = orig {
-                // Search for the captured substring starting from search_from
-                let haystack: Vec<char> = o.chars().collect();
-                let needle: Vec<char> = s.chars().collect();
-                let mut found_from = 0i64;
-                let mut found = false;
-                for start in search_from..=haystack.len().saturating_sub(needle.len()) {
-                    if haystack[start..start + needle.len()] == needle[..] {
-                        found_from = start as i64;
-                        found = true;
-                        break;
-                    }
-                }
-                if found {
-                    (found_from, found_from + needle.len() as i64)
-                } else {
-                    (0i64, s.chars().count() as i64)
-                }
-            } else {
-                (0i64, s.chars().count() as i64)
-            };
-            attrs.insert("from".to_string(), Value::Int(cap_from));
-            attrs.insert("to".to_string(), Value::Int(cap_to));
-            attrs.insert("list".to_string(), Value::array(Vec::new()));
-            attrs.insert("named".to_string(), Value::hash(HashMap::new()));
-            if let Some(o) = orig {
-                attrs.insert("orig".to_string(), Value::str(o.to_string()));
-            }
-            Value::make_instance(Symbol::intern("Match"), attrs)
-        }
-
-        /// Build a Match object from a RegexCaptures, recursively handling subcaptures.
-        fn make_subcap_match(caps: &crate::runtime::RegexCaptures, orig: Option<&str>) -> Value {
-            let search_start = caps.from;
-            let pos_vals: Vec<Value> = caps
-                .positional
-                .iter()
-                .enumerate()
-                .map(|(i, s)| {
-                    // An unmatched optional capture (`(x)?` zero match) renders as Nil.
-                    if caps.positional_nil.get(i) == Some(&true) {
-                        return Value::Nil;
-                    }
-                    // Check if this positional entry is a quantified list
-                    if let Some(Some(qlist)) = caps.positional_quantified.get(i) {
-                        let arr: Vec<Value> = qlist
-                            .iter()
-                            .map(|(text, from, to, subcap)| {
-                                if let Some(sc) = subcap {
-                                    return make_subcap_match(sc, orig);
-                                }
-                                let mut a = HashMap::new();
-                                a.insert("str".to_string(), Value::str(text.clone()));
-                                a.insert("from".to_string(), Value::Int(*from as i64));
-                                a.insert("to".to_string(), Value::Int(*to as i64));
-                                a.insert("list".to_string(), Value::array(Vec::new()));
-                                a.insert("named".to_string(), Value::hash(HashMap::new()));
-                                if let Some(o) = orig {
-                                    a.insert("orig".to_string(), Value::str(o.to_string()));
-                                }
-                                Value::make_instance(Symbol::intern("Match"), a)
-                            })
-                            .collect();
-                        return Value::array(arr);
-                    }
-                    // Recursively build nested Match objects for positional subcaptures
-                    if let Some(Some(subcap)) = caps.positional_subcaps.get(i) {
-                        return make_subcap_match(subcap, orig);
-                    }
-                    make_capture_match(s, orig, search_start)
-                })
-                .collect();
-            let mut sub_named: HashMap<String, Value> = HashMap::new();
-            for (key, values) in &caps.named {
-                let subcaps_for_key = caps.named_subcaps.get(key);
-                let vals: Vec<Value> = values
-                    .iter()
-                    .enumerate()
-                    .map(|(i, s)| {
-                        if let Some(scs) = subcaps_for_key
-                            && let Some(sc) = scs.get(i)
-                        {
-                            return make_subcap_match(sc, orig);
-                        }
-                        make_capture_match(s, orig, search_start)
-                    })
-                    .collect();
-                if vals.len() == 1 && !caps.named_quantified.contains(key) {
-                    sub_named.insert(key.clone(), vals[0].clone());
-                } else {
-                    sub_named.insert(key.clone(), Value::real_array(vals));
-                }
-            }
-            // For quantified named captures that matched zero times, insert empty arrays
-            for qname in &caps.named_quantified {
-                sub_named
-                    .entry(qname.clone())
-                    .or_insert_with(|| Value::real_array(Vec::new()));
-            }
-            // Silent-action captures: hidden `<.foo>` subrule matches (stored under
-            // a marker key in `named_subcaps`) that carry nested captures. They are
-            // absent from `named`/`.hash`, but their action methods must fire, so
-            // build them into a `silent_caps` array for the grammar action walk.
-            // Each subcap's `action_name` carries the rule name to dispatch on.
-            let mut silent_caps_vals: Vec<Value> = Vec::new();
-            for (key, scs) in &caps.named_subcaps {
-                if key.starts_with(crate::runtime::SILENT_ACTION_MARKER_PREFIX) {
-                    for sc in scs {
-                        silent_caps_vals.push(make_subcap_match(sc, orig));
-                    }
-                }
-            }
-            let mut attrs = HashMap::new();
-            attrs.insert("str".to_string(), Value::str(caps.matched.clone()));
-            attrs.insert("from".to_string(), Value::Int(caps.from as i64));
-            attrs.insert("to".to_string(), Value::Int(caps.to as i64));
-            attrs.insert("list".to_string(), Value::array(pos_vals));
-            attrs.insert("named".to_string(), Value::hash(sub_named));
-            if !silent_caps_vals.is_empty() {
-                attrs.insert(
-                    "silent_caps".to_string(),
-                    Value::real_array(silent_caps_vals),
-                );
-            }
-            if let Some(o) = orig {
-                attrs.insert("orig".to_string(), Value::str(o.to_string()));
-            }
-            // Store :sym<> variant name so action dispatch can find it
-            if let Some(ref sym_val) = caps.sym {
-                attrs.insert("sym_variant".to_string(), Value::str(sym_val.clone()));
-            }
-            if let Some(ref action_name) = caps.action_name {
-                attrs.insert("action_name".to_string(), Value::str(action_name.clone()));
-            }
-            if !caps.capture_alias_map.is_empty() {
-                let alias_hash: HashMap<String, Value> = caps
-                    .capture_alias_map
-                    .iter()
-                    .map(|(k, v)| (k.clone(), Value::str(v.clone())))
-                    .collect();
-                attrs.insert("capture_alias_map".to_string(), Value::hash(alias_hash));
-            }
-            Value::make_instance(Symbol::intern("Match"), attrs)
-        }
-
-        let mut attrs = HashMap::new();
-        attrs.insert("str".to_string(), Value::str(matched));
-        attrs.insert("from".to_string(), Value::Int(from));
-        attrs.insert("to".to_string(), Value::Int(to));
-        if let Some(o) = orig {
-            attrs.insert("orig".to_string(), Value::str(o.to_string()));
-        }
-        let search_start = from as usize;
-        let caps: Vec<Value> = positional
-            .iter()
-            .enumerate()
-            .map(|(i, s)| {
-                // An unmatched optional capture (`(x)?` zero match) renders as Nil.
-                if positional_nil.get(i) == Some(&true) {
-                    return Value::Nil;
-                }
-                // Check if this positional entry is a quantified list
-                if let Some(Some(qlist)) = positional_quantified.get(i) {
-                    let arr: Vec<Value> = qlist
-                        .iter()
-                        .map(|(text, qfrom, qto, subcap)| {
-                            if let Some(sc) = subcap {
-                                return make_subcap_match(sc, orig);
-                            }
-                            let mut a = HashMap::new();
-                            a.insert("str".to_string(), Value::str(text.clone()));
-                            a.insert("from".to_string(), Value::Int(*qfrom as i64));
-                            a.insert("to".to_string(), Value::Int(*qto as i64));
-                            a.insert("list".to_string(), Value::array(Vec::new()));
-                            a.insert("named".to_string(), Value::hash(HashMap::new()));
-                            if let Some(o) = orig {
-                                a.insert("orig".to_string(), Value::str(o.to_string()));
-                            }
-                            Value::make_instance(Symbol::intern("Match"), a)
-                        })
-                        .collect();
-                    return Value::array(arr);
-                }
-                // If there are nested subcaptures for this positional capture,
-                // build a full Match object with subcaptures
-                if let Some(Some(subcap)) = positional_subcaps.get(i) {
-                    return make_subcap_match(subcap, orig);
-                }
-                make_capture_match(s, orig, search_start)
-            })
-            .collect();
-        attrs.insert("list".to_string(), Value::array(caps));
-        let mut named_caps_map: HashMap<String, Value> = HashMap::new();
-        for (key, values) in named {
-            let subcaps_for_key = named_subcaps.get(key);
-            let vals: Vec<Value> = values
-                .iter()
-                .enumerate()
-                .map(|(i, s)| {
-                    if let Some(scs) = subcaps_for_key
-                        && let Some(sc) = scs.get(i)
-                    {
-                        return make_subcap_match(sc, orig);
-                    }
-                    make_capture_match(s, orig, search_start)
-                })
-                .collect();
-            if vals.len() == 1 && !named_quantified.contains(key) {
-                named_caps_map.insert(key.clone(), vals[0].clone());
-            } else {
-                named_caps_map.insert(key.clone(), Value::real_array(vals));
-            }
-        }
-        // For quantified named captures that matched zero times, insert empty arrays
-        for qname in named_quantified {
-            named_caps_map
-                .entry(qname.clone())
-                .or_insert_with(|| Value::real_array(Vec::new()));
-        }
-        attrs.insert("named".to_string(), Value::hash(named_caps_map));
-        // Silent-action captures: hidden `<.foo>` subrule matches carrying nested
-        // captures (see make_subcap_match). Build them into a `silent_caps` array
-        // so the grammar action walk fires their (and descendants') actions; they
-        // are deliberately absent from `named`/`.hash`.
-        let mut silent_caps_vals: Vec<Value> = Vec::new();
-        for (key, scs) in named_subcaps {
-            if key.starts_with(crate::runtime::SILENT_ACTION_MARKER_PREFIX) {
-                for sc in scs {
-                    silent_caps_vals.push(make_subcap_match(sc, orig));
-                }
-            }
-        }
-        if !silent_caps_vals.is_empty() {
-            attrs.insert(
-                "silent_caps".to_string(),
-                Value::real_array(silent_caps_vals),
-            );
-        }
-        Value::make_instance(Symbol::intern("Match"), attrs)
-    }
-
-    pub(crate) fn version_strip_trailing_zeros(parts: &[VersionPart]) -> Vec<VersionPart> {
-        let mut v: Vec<VersionPart> = parts.to_vec();
-        while matches!(v.last(), Some(VersionPart::Num(0))) {
-            v.pop();
-        }
-        if v.is_empty() {
-            vec![VersionPart::Num(0)]
-        } else {
-            v
-        }
-    }
-
-    pub(crate) fn is_range(&self) -> bool {
-        matches!(
-            self,
-            Value::Range(_, _)
-                | Value::RangeExcl(_, _)
-                | Value::RangeExclStart(_, _)
-                | Value::RangeExclBoth(_, _)
-                | Value::GenericRange { .. }
-        )
-    }
-
-    /// Check if this value is a numeric type (Int, Num, Rat, FatRat, BigInt).
-    /// Returns the inner items if this value is an Array, Seq, or Slip.
-    pub(crate) fn as_list_items(&self) -> Option<&[Value]> {
-        match self {
-            Value::Array(items, _) => Some(&items.items[..]),
-            Value::Seq(items) | Value::Slip(items) => Some(&items[..]),
-            _ => None,
-        }
-    }
-
-    pub(crate) fn is_numeric(&self) -> bool {
-        matches!(
-            self,
-            Value::Int(_)
-                | Value::BigInt(_)
-                | Value::Num(_)
-                | Value::Rat(_, _)
-                | Value::FatRat(_, _)
-                | Value::BigRat(_, _)
-                | Value::Whatever
-        )
-    }
-
-    /// Convert a numeric value to f64.
-    pub(crate) fn to_f64(&self) -> f64 {
-        match self {
-            // Phase 2 element container: numify the inner value transparently
-            // if a `:=`-bound element's cell leaks into a numeric context.
-            Value::ContainerRef(cell) => cell.lock().unwrap().to_f64(),
-            Value::Int(i) => *i as f64,
-            Value::BigInt(n) => n.to_f64().unwrap_or(0.0),
-            Value::Num(f) => *f,
-            Value::Rat(n, d) => {
-                if *d != 0 {
-                    *n as f64 / *d as f64
-                } else if *n == 0 {
-                    f64::NAN
-                } else if *n > 0 {
-                    f64::INFINITY
-                } else {
-                    f64::NEG_INFINITY
-                }
-            }
-            Value::FatRat(n, d) => {
-                if *d != 0 {
-                    *n as f64 / *d as f64
-                } else if *n == 0 {
-                    f64::NAN
-                } else if *n > 0 {
-                    f64::INFINITY
-                } else {
-                    f64::NEG_INFINITY
-                }
-            }
-            Value::BigRat(n, d) => {
-                if !d.is_zero() {
-                    n.to_f64().unwrap_or(0.0) / d.to_f64().unwrap_or(1.0)
-                } else if n.is_zero() {
-                    f64::NAN
-                } else if n.is_positive() {
-                    f64::INFINITY
-                } else {
-                    f64::NEG_INFINITY
-                }
-            }
-            Value::Bool(b) => {
-                if *b {
-                    1.0
-                } else {
-                    0.0
-                }
-            }
-            Value::Whatever => f64::INFINITY,
-            Value::Str(s) => s.trim().parse::<f64>().unwrap_or(0.0),
-            Value::Array(items, ..) => items.len() as f64,
-            Value::Hash(map) => map.len() as f64,
-            Value::Instance {
-                class_name,
-                attributes,
-                ..
-            } if class_name == "Instant" || class_name == "Duration" => attributes
-                .as_map()
-                .get("value")
-                .map(|v| v.to_f64())
-                .unwrap_or(0.0),
-            // A subclass of native Int (e.g. `class Foo is Int`) carries its
-            // integer payload in the reserved `__mutsu_int_value` attribute.
-            Value::Instance { attributes, .. } if attributes.contains_key("__mutsu_int_value") => {
-                attributes
-                    .as_map()
-                    .get("__mutsu_int_value")
-                    .map(|v| v.to_f64())
-                    .unwrap_or(0.0)
-            }
-            // Match coerces to Numeric via its matched string
-            Value::Instance {
-                class_name,
-                attributes,
-                ..
-            } if class_name == "Match" => attributes
-                .as_map()
-                .get("str")
-                .map(|v| v.to_string_value().trim().parse::<f64>().unwrap_or(0.0))
-                .unwrap_or(0.0),
-            _ => 0.0,
-        }
-    }
-
-    /// Convert a Value to a num_bigint::BigInt for arbitrary-precision arithmetic.
-    pub(crate) fn to_bigint(&self) -> NumBigInt {
-        match self {
-            Value::Int(i) => NumBigInt::from(*i),
-            Value::BigInt(n) => (**n).clone(),
-            Value::Num(f) => NumBigInt::from(*f as i64),
-            Value::Rat(n, d) => {
-                if *d != 0 {
-                    NumBigInt::from(n / d)
-                } else {
-                    NumBigInt::from(0)
-                }
-            }
-            Value::BigRat(n, d) => {
-                if !d.is_zero() {
-                    n.as_ref() / d.as_ref()
-                } else {
-                    NumBigInt::from(0)
-                }
-            }
-            Value::Str(s) => s
-                .parse::<NumBigInt>()
-                .unwrap_or_else(|_| NumBigInt::from(0)),
-            // A subclass of native Int (e.g. `class Foo is Int`) carries its
-            // integer payload in the reserved `__mutsu_int_value` attribute.
-            Value::Instance { attributes, .. } => attributes
-                .as_map()
-                .get("__mutsu_int_value")
-                .map(|v| v.to_bigint())
-                .unwrap_or_else(|| NumBigInt::from(0)),
-            Value::Mixin(inner, _) => inner.to_bigint(),
-            _ => NumBigInt::from(0),
-        }
-    }
-
-    /// Create a Value from a BigInt, normalizing to Int(i64) when possible.
-    pub(crate) fn from_bigint(n: NumBigInt) -> Value {
-        if let Some(i) = n.to_i64() {
-            Value::Int(i)
-        } else {
-            Value::bigint(n)
-        }
-    }
 }
 
 // Compile-time assertions that Value is Send + Sync
